==> R/03-typology_1nga.Rmd <==
# Typology of Farm Holdings

This section is an exploration of smallholdings and smallholder farmers (definition and characteristics) in Nigeria and Ghana.

This study combines micro data for 4 countries across multiple survey years. To expedite all data reconciliation and preparation steps we choose to rely as much as possible on FAO/RIGA cross-country harmonized database of income aggregates. These aggregates are used to estimate both farm and crop sales and farm vs. non-farm income. Generally in the RIGA approach the method is (see @riga2007doco):

- All aggregates are estimated at the household level.  
- All income and expenditures are **annualized**.  
- All income components are **net of costs**.  
- Purchases and sales of **durable goods**, investments and windfall gains are excluded from household income and expenditure calculations since these are not transactions undertaken regularly by households and can result in the significant over- or under-stating of permanent income.  
- All aggregates are reported in **local currency units (LCU)**.  
- For each survey, only the **rural** sample is used (this is not consustent across all surveys)  
- An outlier check is imposed after the construction of all income aggregates to deal with extreme income shares that arise following the aggregation of all income components. After estimating the shares of the seven principal income categories (agricultural wages, non-agricultural wages, crop, livestock, self employment, transfer and other) **observations whose share is greater than or less than 3 (indicating a percentage share of +/- 300%) are dropped from the aggregate**.  

However because RIGA income aggregates are *net of costs*, we use intermediary components instead. Details are also shown in the table below.

- $cropsales$ derived from $totcropsold$
- $cropvalue$ derived from $totcropprod$
- $cropsales\_sh$ (but can also be derived from $shcropsold$ and $shagprodsold$)
  $$ cropsales\_sh = \frac{cropsales}{cropvalue} $$ 
- $agsales$ derived from $totagsold$
- $aggross = totagprod + farmrntimp + agr\_wge (+ selfimp1)$
- $naggross = transfersgrossimp + otherincimp + nonfarmrntimp + nonagr\_wge + selfemp$
- $totgross = aggross + naggross$  
- $aggross\_sh$  
  $$ aggross\_sh = \frac{aggross}{totgross} $$  

RIGA does not provide annual cultivated area, instead we have measures of land owned, land rented out, and land under sharecropping.

## Nigeria

```{r nga-ghs, eval=FALSE}

library(foreign)
library(data.table)
library(survey)
library(tables)
library(stringr)

setwd("~/Projects/2017-agra-aasr")
load("./tmp/2017-agra-aasr.RData")

#####################################################################################
# Helper - Combine and format results from survey::svyby()
svyCrossTab <- function(formula, by, design, quantiles=c(.25,.5,.75)) {
  
  # Pass a list of formulas (list(~var1, ~var2, ...))
  # Then use tables::tabular() to format pivot tables
  
  require(survey)
  require(data.table)
  require(stringr)
  
  # Reshape means
  x <- lapply(formula, function(x) {
    m <- svyby(x, by, design, svymean, na.rm=TRUE)
    m <- as.data.table(ftable(m))
    return(m)
  })
  x <- rbindlist(x)
  setnames(x, c("by", "stat", "var", "Mean"))
  x[stat=="svymean", stat := "est."]
  x[stat=="SE", stat := "std. err."]
  
  # Reshape quantiles
  y <- lapply(formula, function(x) {
    m <- svyby(x, by, design, svyquantile, quantiles, ci=TRUE, na.rm=TRUE)
    m <- as.data.table(ftable(m))
    m[, var := as.character(x)[2]]
    return(m)
  })
  y <- rbindlist(y)  
  setnames(y, c("by", "stat", "qtl", "value", "var"))
  y[stat=="svyquantile", stat := "est."]
  y[stat=="SE", stat := "std. err."]
  y[, qtl := paste0("Q", as.numeric(as.character(qtl))*100)]
  y <- dcast(var+by+stat~qtl, data=y)
  
  setkey(x, var, by, stat)
  setkey(y, var, by, stat)
  x <- x[y]
 
  # Add optional labels
  if(!is.null(names(formula))) for (i in seq_along(var)) x[var==as.character(formula[[i]])[2], var := names(formula)[i]]
  
  var.levels <- sapply(formula, function(x) as.character(x)[2])
  var.levels <- str_replace(var.levels, fixed("I("), fixed("("))
  x[, var := factor(str_replace(var, fixed("I("), fixed("(")), levels=var.levels)]
  x[, stat := factor(as.character(stat), ordered=T)]
  setcolorder(x, c(3,1,2,4:ncol(x)))
  setnames(x, 1:3, c("Variable", "By", "Stat"))
  return(x)
}

#####################################################################################
# Helper - Combine multiple svyhist(), svysmooth(), and svycdf() plots
svyCrossPlot <- function(formula, design, breaks="Sturges", bandwidth=NULL, ...) {
  
  require(survey)
  
  n <- class9
  par(mfrow=c(length(n), 2))
  for (i in n) {
    svyhist(formula, subset(design, class9==i), breaks=breaks,
      ylab="Density", main=i, ...)
    lines(svysmooth(formula, subset(design, class9==i), bandwidth=bandwidth), 
      lwd=1, col=40)
    plot(svycdf(formula, subset(design, class9==i)), 
      ylab="Est. CDF", main=i, ...)
    abline(h=0.5, lty=2, col=40)
    abline(h=0, lty=2, col="grey70")
    abline(h=1, lty=2, col="grey70")
  }
}
#####################################################################################

# Load and copy constructed vars from the 2016 NGA Segmentation (GHS)
e1 <- new.env()
setwd("~/Google Drive/2016-BMGF-Segmentation/NGA")
load("./temp/nga_bmgf_seg_1.RData", envir=e1)
setwd("~/Projects/2017-agra-aasr")

ghs <- e1$ghs
ghs.farm <- e1$farm
ghs.geo <- e1$geo12
ghs.inc <- e1$inc12
ghs.ppp <- e1$ppp12
rm(e1)

# Verify all income aggregates in `ghs.inc` (received from Cleo/IFPRI)
# Cleo suggested using crop/fish/livestock and by-products sold/gift/lost/reimbursed
# `cropsold`, `agbyprodincimp`, `reimbursmt_n`, `gift_n` in `Income2.dta`
# Note that last 3 vars have negative sign, need to correct
summary(ghs.inc$hh)
#  Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# 10000   90040  190000  183200  280000  370000
summary(ghs$hhid)
summary(ghs.ppp$hhid)
# => `hh` match `hhid`

vars <- c("cropsold", "croplost", "cropownimp",
  "reimbursmt_n", "lost_n", "gift_n", "self",
  "fishincimp", "agr_wge", "nonagr_wge", "other",
  "farmrntimp", "agrentimp", "privtransfer", "pubtransfer")
ghs.inc[, lapply(names(.SD), function(x) boxplot(.SD[[x]], main=x)), .SDcols=vars]

# Reconstruct all value vars in `ghs`
# `livstrevimp`, `livstbyprod`, `fishincimp` have been corrected in separate files
# Use `ghs$lstocksold` instead of `ghs.inc$livstrevimp`
# Use `ghs$fishsold` instead of `ghs.inc$fishincimp`
vars.val <- c("cropsales", "cropvalue", "agsales", "agsalespc", "aggross", "totgross", "cropsales_sh", "aggross_sh")
ghs[, (unique(c(vars, vars.val))) := NULL]

setkey(ghs, hhid)
setkey(ghs.inc, hh)
setkey(ghs.farm, hhid)
ghs[ghs.farm, croparea := croparea]

ghs[ghs.inc, cropsales := rowSums(.SD[, .(cropsold, agbyprodincimp, -reimbursmt_n, -gift_n)], na.rm=T)]
summary(ghs$cropsales)
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's
#    0       0       0   37910   18500 3068000      19

ghs[ghs.inc, cropvalue := rowSums(.SD[, .(cropsales, croplost, cropownimp, -lost_n)], na.rm=T)]
summary(ghs$cropvalue)
# Min.   1st Qu.    Median      Mean   3rd Qu.      Max.      NA's
#    0         0     68370  14990000  23400000 161000000        19

ghs[, agsales := rowSums(.SD[, .(cropsales, lstocksold, fishsold)], na.rm=T)]
summary(ghs$agsales)
# Min.  1st Qu.   Median     Mean  3rd Qu.     Max.
#    0        0        0   240900    63420 90620000

ghs[, agsalespc := agsales/equiv]
ghs[ghs.inc, aggross := rowSums(.SD[, .(agsales, agr_wge, farmrntimp, agrentimp, cropownimp)], na.rm=T)]
summary(ghs$aggross)
# Min.   1st Qu.    Median      Mean   3rd Qu.      Max.      NA's
#    0         0    150900  15200000  23450000 163100000        19

ghs[ghs.inc, totgross := rowSums(.SD[, .(aggross, nonagr_wge, other, privtransfer, pubtransfer, self)], na.rm=T)]
summary(ghs$totgross)
# Min.   1st Qu.    Median      Mean   3rd Qu.      Max.      NA's
#    0    287500   1525000  15630000  23620000 163100000        19

ghs[, noaggross := rowSums(.SD[, .(totgross, -aggross)], na.rm=T)]
summary(ghs$noaggross)
# Min.  1st Qu.   Median     Mean  3rd Qu.     Max.
#    0     2075   172700   419800   499400 10150000

ghs[, aggross_sh := aggross/totgross]
ghs[, noaggross_sh := 1 - aggross_sh]
ghs[, cropsales_sh := cropsales/cropvalue]

# Fix Inf values
ghs[agsalespc==Inf, agsalespc := NA]

# Define a farm holding as a household matching any of these conditions:
# - has any `farmarea` or `croparea`
# - has any `cropvalue` or `lstocksold` or `fishsold`
# - has any `TLU_total`
# - has any ag revenue `aggross`
ghs[, boxplot(TLU_total)]
ghs[, farm_hh := FALSE]
ghs[farmarea>0 | croparea>0 | cropvalue>0 | lstocksold>0 | fishsold>0 | TLU_total>0 | aggross>0,
  farm_hh := TRUE]
ghs[, summary(farm_hh)]
#    Mode   FALSE    TRUE    NA's
# logical    1537    3343       0

# Which farms in `ghs` are not accounted for in `ghs.farm`?
ghs.farm[!ghs[farm_hh==T], .N]
# [1] 116
ghs[farm_hh==T][!ghs.farm, .N]
# [1] 278
ghs.farm[!ghs[farm_hh==T], croparea]
# => all 0 or NA
ghs[farm_hh==T][!ghs.farm, cropsales]
# => some crop sales

# SHF 4ha dummy `croparea_4ha` (2-class)
# Merge `croparea` into `ghs` (`hhid` are unique)
ghs[, croparea_4ha := NULL]
ghs[farm_hh==T, croparea_4ha := factor(croparea <= 4, levels=c(T, F), labels=c("<= 4 ha", "> 4 ha"))]
summary(ghs$croparea_4ha)
# <= 4 ha  > 4 ha    NA's
#    2840      94    1946

ghs[croparea==0, summary(croparea_4ha)]
# <= 4 ha  > 4 ha    NA's
#     240       0       7
# => need to re-run results without these 240
ghs[croparea==0, summary(cropsales)]
ghs[croparea==0, summary(TLU_total)]

# SHF 4ha dummy `croparea_2ha` (2-class)
ghs[, croparea_2ha := NULL]
ghs[farm_hh==T, croparea_2ha := factor(croparea <= 2, levels=c(T, F), labels=c("<= 2 ha", "> 2 ha"))]
summary(ghs$croparea_2ha)
# <= 2 ha  > 2 ha    NA's
#    2616     318    1946

# Remove farms with no croparea from this classification
ghs[croparea==0, croparea_2ha := NA]
summary(ghs$croparea_2ha)
# <= 2 ha  > 2 ha    NA's 
#    2376     318    2186 

# SHF 4ha dummy `croparea_clas` (3-class)
ghs[, croparea_clas := NULL]
ghs[farm_hh==T, croparea_clas := cut(croparea, c(-1,2,4,25), labels=c("<= 2 ha", "2-4 ha", "> 4 ha"), ordered=T, right=F)]
summary(ghs$croparea_clas)
# <= 2 ha  2-4 ha  > 4 ha    NA's
#    2616     224      94    1946

# Remove farms with no croparea from this classification
ghs[croparea==0, croparea_clas := NA]
summary(ghs$croparea_clas)
# <= 2 ha  2-4 ha  > 4 ha    NA's 
#    2376     224      94    2186 

# Crop commercialization dummy `cropsales_any` (2-class)
ghs[, cropsales_any := NULL]
ghs[farm_hh==T, cropsales_any := factor(cropsales_sh > 0, levels=c(T, F), labels=c("sales", "no sales"))]
summary(ghs$cropsales_any)
# sales no sales     NA's
#  1868     1234     1778

# Crop commercialization dummy `cropsales_clas` (3-class)
ghs[, cropsales_clas := NULL]
ghs[farm_hh==T, cropsales_clas := cut(cropsales_sh, c(-1,.05,.5,1.1), labels=c("LoC", "MeC", "HiC"), ordered=T, right=F)]
summary(ghs$cropsales_clas)
#  LoC  MeC  HiC NA's
# 2559  239  304 1778

# Income diversification dummy `noaggross_clas` (3-class)
ghs[, noaggross_clas := NULL]
ghs[farm_hh==T, noaggross_clas := cut(noaggross_sh, c(-1,1/3,2/3,1.1), labels=c("LoD", "MeD", "HiD"), ordered=T, right=F)]
summary(ghs$noaggross_clas)
#  LoD  MeD  HiD NA's
# 2528  172  608 1572

class9 <- c("LoCLoD", "LoCMeD", "LoCHiD",
  "MeCLoD", "MeCMeD", "MeCHiD", "HiCLoD", "HiCMeD", "HiCHiD")

# Classification dummy `` (9-class)
ghs[, class9 := NULL]
ghs[, class9 := factor(paste0(cropsales_clas, noaggross_clas), levels=class9)]
summary(ghs$class9)
# LoCLoD LoCMeD LoCHiD MeCLoD MeCMeD MeCHiD HiCLoD HiCMeD HiCHiD   NA's
#   2220     54    285    121     34     84    125     70    109   1778

# Merge new dummay vars into `ghs.farm` and `ghs.ppp`
setkey(ghs, hhid)
setkey(ghs.farm, hhid)
setkey(ghs.ppp, hhid)
setkey(ghs.geo, hhid)

ghs.farm[, c("croparea_2ha", "croparea_4ha",
  "cropsales_sh", "cropsales_clas",
  "noaggross_clas", "class9") := NULL]
ghs.ppp[, c("croparea_2ha", "croparea_4ha",
  "cropsales_sh", "cropsales_clas",
  "noaggross_clas", "class9") := NULL]
ghs.geo[, c("croparea_2ha", "croparea_4ha",
  "cropsales_sh", "cropsales_clas",
  "noaggross_clas", "class9", "wt_wave2", "zone", "sector", "rural") := NULL]

ghs.farm[ghs, `:=`(
  croparea_2ha = croparea_2ha,
  croparea_4ha = croparea_4ha,
  cropsales_sh = cropsales_sh,
  cropsales_clas = cropsales_clas,
  noaggross_clas = noaggross_clas,
  class9 = class9
)]

ghs.ppp[ghs, `:=`(
  zone = zone,
  croparea_2ha = croparea_2ha,
  croparea_4ha = croparea_4ha,
  cropsales_sh = cropsales_sh,
  cropsales_clas = cropsales_clas,
  noaggross_clas = noaggross_clas,
  class9 = class9
)]

ghs.geo[ghs, `:=`(
  croparea_2ha = croparea_2ha,
  croparea_4ha = croparea_4ha,
  cropsales_sh = cropsales_sh,
  cropsales_clas = cropsales_clas,
  noaggross_clas = noaggross_clas,
  class9 = class9,
  wt_wave2 = wt_wave2,
  zone = zone,
  rural = rural
  )]

# Add a few extra geo vars to `ghs` to plot
ghs[ghs.geo, `:=`(
  dist_market = dist_market,
  dist_road2 = dist_road2
  )]

##########################################
# Define 2-stage stratified survey design (use most recent GHS wave only)
ghs.svy <- svydesign(~ea+hhid, strata=~zone+rural, weights=~wt_wave2,
  data=ghs[!is.na(wt_wave2)], nest=T)
ghs.farm.svy <- svydesign(~ea+hhid, strata=~zone+sector, weights=~wt_wave2,
  data=ghs.farm[!is.na(wt_wave2)], nest=T)
ghs.ppp.svy <- svydesign(~hhid, strata=~zone+rural, weights=~weight,
  data=ghs.ppp[!is.na(weight)], nest=T)
ghs.geo.svy <- svydesign(~hhid, strata=~zone+rural, weights=~wt_wave2,
  data=ghs.geo[!is.na(wt_wave2)], nest=T)

summary(ghs.svy)
# Stratified 2 - level Cluster Sampling design (with replacement)
# With (481, 4867) clusters.
# svydesign(~ea + hhid, strata = ~zone + rural, weights = ~wt_wave2,
#     data = ghs[!is.na(wt_wave2)], nest = T)
# Probabilities:
#      Min.   1st Qu.    Median      Mean   3rd Qu.      Max.
# 0.0000239 0.0001235 0.0001892 0.0002150 0.0002699 0.0014700
# First-level Stratum Sizes:
#            north central north east north west south east south south south west
# obs                  798        754        890        791         786        848
# design.PSU            79         74         85         77          79         87
# actual.PSU            79         74         85         77          79         87

# Subset survey sample to farms
ghs.svy.farm <- subset(ghs.svy, farm_hh==T)
# Subset survey sample to only SHF (note that they exclude farms with no `croparea`)
ghs.svy.shf <- subset(ghs.svy, croparea_4ha=="<= 4 ha")
ghs.farm.svy.shf <- subset(ghs.farm.svy, croparea_4ha=="<= 4 ha")
ghs.ppp.svy.shf <- subset(ghs.ppp.svy, croparea_4ha=="<= 4 ha")
ghs.geo.svy.shf <- subset(ghs.geo.svy, croparea_4ha=="<= 4 ha")

# Verify sampling design (e.g. total population)
(svytotal(~hhsize, ghs.svy, na.rm=T))/1E6
#         total      SE
# hhsize 187.88 7601475

svytotal(~farm_hh, ghs.svy, na.rm=T)/1E6
#               total      SE
# farm_hhFALSE 13.480 1259455
# farm_hhTRUE  19.256  712970

svymean(~farm_hh, ghs.svy, na.rm=T)
#                 mean     SE
# farm_hhFALSE 0.41177 0.0249
# farm_hhTRUE  0.58823 0.0249

svymean(~croparea_4ha, ghs.svy.farm, na.rm=T)
#                         mean     SE
# croparea_4ha<= 4 ha 0.971845 0.0034
# croparea_4ha> 4 ha  0.028155 0.0034

svytotal(~hhsize, ghs.svy.shf, na.rm=T)
#            total      SE
# hhsize 106349144 4497228

svymean(~cropsales_sh, ghs.svy.farm, na.rm=T)
#                mean     SE
# cropsales_sh 0.1141 0.0087

svyby(~cropsales_sh, ~croparea_4ha, ghs.svy.farm, svymean, na.rm=T)
#         croparea_4ha cropsales_sh         se
# <= 4 ha      <= 4 ha    0.1220143 0.00648625
# > 4 ha        > 4 ha    0.1768049 0.04914111

svyby(~cropsales_sh, ~croparea_4ha, ghs.svy.farm, svyquantile, quantiles=.5, ci=T, na.rm=T)*100
#         croparea_4ha cropsales_sh          se
# <= 4 ha           NA    0.0771759 0.009096591
# > 4 ha            NA    0.2983541 0.297538378

svyby(~agsalespc, ~croparea_4ha, ghs.svy.farm, svymean, na.rm=T)
#         croparea_4ha agsalespc       se
# <= 4 ha      <= 4 ha   87756.2 13964.07
# > 4 ha        > 4 ha  140465.3 44667.0

svymean(~I(noaggross_sh*100), subset(ghs.svy.farm, !is.na(croparea)), na.rm=T)
#                         mean     SE
# I(noaggross_sh * 100) 22.044 0.0127
# I(noaggross_sh * 100) 17.446 1.1725

svyquantile(~I(noaggross_sh*100), ghs.svy.farm, .5, na.rm=T)
#                             0.5
# I(noaggross_sh * 100) 0.8265929

svyby(~I(noaggross_sh*100), ~croparea_4ha, ghs.svy.farm, svymean, na.rm=T)
#         croparea_4ha I(noaggross_sh * 100)       se
# <= 4 ha      <= 4 ha              17.58628 1.190880
# > 4 ha        > 4 ha              12.53712 3.285989

svyby(~I(noaggross_sh*100), ~croparea_4ha, ghs.svy.farm, svyquantile, .5, ci=T, na.rm=T)
#         croparea_4ha I(noaggross_sh * 100)         se
# <= 4 ha      <= 4 ha             0.6300151 0.08870994
# > 4 ha        > 4 ha             0.3270542 0.12108961

# Other farms statistics
tmp <- ghs[!is.na(wt_wave2) & farm_hh==T, .(
  f2malehh_tot=sum(femhead*wt_wave2, na.rm=T),
  malehh_tot=sum(-(femhead-1)*wt_wave2, na.rm=T),
  femalehh=weightedMean(femhead, wt_wave2, na.rm=T),
  hasland=weighted.mean(farmarea>0, wt_wave2, na.rm=T),
  farmarea=weightedMean(hectares_1_hh_farmer, wt_wave2, na.rm=T),
  farmarea_med=weightedMedian(hectares_1_hh_farmer, wt_wave2, na.rm=T)
), keyby=croparea <= 4]

tmp <- farm[, .(
  farmslive_tot=sum(farmslive*wt_wave2, na.rm=T),
  farmslive=weightedMean(farmslive, wt_wave2, na.rm=T),
  intens_idx2=weightedMean(intens_idx2, wt_wave2, na.rm=T),
  intens_idx2_med=weightedMedian(intens_idx2, wt_wave2, na.rm=T)
), keyby=croparea <= 4]

# Livestock
tmp <- ghs[!is.na(wt_wave2) & totgross > 0, .(
  TLU_total=sum(TLU_total, wt_wave2, na.rm=T),
  TLU_hh=weightedMean(TLU_total, wt_wave2, na.rm=T),
  TLU_hh_imp=weightedMean(TLU_total_imp, wt_wave2, na.rm=T)
), keyby=croparea <= 4][, wt_wave2 := NULL]

# Input uses
tmp <- farm[, lapply(.SD, weightedMean, wt_wave2, na.rm=T),
  .SDcols=c("wt_wave2", "parcels", "sh_chempest", "sh_chemherb", "sh_fertorg",
    "sh_fertinorg", "sh_seedp", "tractorown_farms"),
  keyby=croparea <= 4][, wt_wave2 := NULL]

# Poverty
vars <- c("weight", names(ppp12)[names(ppp12) %like% "ppp"])
tmp <- ppp12[, lapply(.SD, weightedMean, weight, na.rm=T), .SDcols=vars,
  keyby=croparea <= 4][, weight := NULL]

# CROP SALES


# FOOD EXPENDITURES
m <- svymean(~I(100*foodexp_ppp_m/rowSums(cbind(foodexp_ppp_m, nfoodexp_ppp_m), na.rm=T)), ghs.ppp.svy, na.rm=T)
#                                                                                   mean     SE
# I(100 * foodexp_ppp_m/rowSums(cbind(foodexp_ppp_m, nfoodexp_ppp_m), na.rm = T)) 73.394 0.2999

plot(m)

svyquantile(~I(100*foodexp_ppp_m/rowSums(cbind(foodexp_ppp_m, nfoodexp_ppp_m), na.rm=T)), ghs.ppp.svy, .5, ci=T, na.rm=T)
#                                                                                      0.5
# I(100 * foodexp_ppp_m/rowSums(cbind(foodexp_ppp_m, nfoodexp_ppp_m), na.rm = T)) 76.27067

# Save workspace
rm(i, x, tmp)
save.image("./tmp/2017-agra-aasr.RData")

```

Relevant sources of household-level micro records for Nigeria include:

* 2010 and 2012 Nigeria General Household Survey Panels
* 2003 and 2009/10 Harmonized Nigeria Living Standard Surveys
* DfID's sponsored agricultural survey (1,300 hhlds) (contact Abt Associates)
* Fadama III 2009 Baseline and 2012 Midline Surveys and Fadama's Agricultural Production Surveys (APS) (still not publicly available, see e.g. @girei2017fadama)
* 2010/11 National Agricultural Sample Survey (NASS) (only a PDF is available)
* 2006 Nigeria Private Farmer Crop Sample Survey
* 2006 Nigeria Private Farmer Livestock Poultry Sample Survey

Note that all intermediary *value* variables (income and expenses) have been generated by IFPRI using [FAO RIGA](http://www.fao.org/economic/riga/riga-database/en/) guidelines. 

> Using these variables we define a **farm holding** as a household matching any of the following conditions:
> 
> * has any $farmarea$ or $croparea$
> * has any $cropvalue$ or $lstocksold$ or $fishsold$
> * owns any livestock $TLU\_total$
> * collects any other revenue from farming $aggross$
> 
> Within the sub-sample of *farm households* we look at the distribution of households across 3 key ("threshold") variables:
> 
> * Cultivated/operated area $croparea$ (using breaks at **2 ha** and/or **4 ha** and **excluding farm households with no $croparea$**)
> * Crop commercialization $cropsales\_sh$ (using breaks at **5%** and **50%** of crop production)
> * Share of off-farm income in total hhld income $naggross\_sh$ (using breaks at **1/3** and **2/3** of total gross income)

The resulting classes of farm households are labeled as follows:  

```{r tabclas, results="asis"}

tmp <- factor(c("low", "medium", "high"), levels=c("low", "medium", "high"))
tmp <- data.frame(clas=class9, C=rep(tmp, each=3), D=rep(tmp, 3))

table_options(HTMLcaption="(#tab:tabclas) Household Classification")
html(tabular((`Income Diversification`=D)~(`Crop Commercialization`=C)*Heading()*identity*Heading()*(as.character(clas)), 
  data=tmp))

```


### An Aside on Income Diversification

Initial results from the 2 Nigeria GHS panels tend to show very low income diversification among rural farm households. We propose to check these results against income and expenditure aggregates derived by FAO RIGA. RIGA documentation for Nigeria indicates the following aggregate income shares [@riga2016nga] (among rural households only):

```{r riga-net}

tmp <- fread("../data/tabula-Nigeria13-Components_of_Income_Aggregate.csv")
kable(tmp[, .SD, .SDcols=c(1,3:8)], split.table=Inf,
  justify="crrrrrr", round=2, format.args=list(big.mark=","), keep.trailing.zeros=T,
  caption="RIGA Income Aggregates, NGHS 2012/2013 (Rural Households)")

```

All income aggregates are *net* of expenses. For this study however *gross* crop and livestock sales and *gross* farm and non-farm incomes would be preferable. These are derived in the codes below [in progress].

```{r riga-gross}

# Load RIGA components



```

### Key Results

The following 3 tables show the count of observations in each class for farm households with cultivated area <= 2ha, between 2 and 4ha, and <= 4ha.

```{r ngacounts, results="asis"}

table_options(HTMLcaption="(#tab:countsa) Count of Observations across Classes of Farm Households")
html(tabular(Heading()*noaggross_clas+1~Format(digits=0, big.mark=",")*Heading()*croparea_clas*Heading()*cropsales_clas+1,
  data=ghs))

table_options(HTMLcaption="(#tab:countsb) Count of Observations across Classes of Farm Households (below/above 2 ha)")
html(tabular(Heading()*noaggross_clas+1~Format(digits=0, big.mark=",")*Heading()*croparea_2ha*Heading()*cropsales_clas+1,
  data=ghs))

table_options(HTMLcaption="(#tab:countsc) Count of Observations across Classes of Farm Households (below/above 4 ha)")
html(tabular(Heading()*noaggross_clas+1~Format(digits=0, big.mark=",")*Heading()*croparea_4ha*Heading()*cropsales_clas+1,
  data=ghs))

```

In the next tables we limit categories to farm households **below/above 4ha** (as this approach produces classes with a larger count of observations). Using data from the 2012 Nigeria General Household Survey (GHS) we estimate the proportions of farm households across all categories and the distributions of other key household characteristics (household structure, assets, farm input uses, etc.).

Estimated proportions of farm households across classes and categories are shown in the next graph and table.

```{r nga-fig2a, fig.cap="Est. Proportions of Farm Holdings across Categories"}

par(mfrow=c(1,3), mar=c(3,3,0,.25))
plot(svytable(~croparea_4ha+cropsales_clas, ghs.svy.shf), main=NA, las=1, col=viridis(3),
  ylab="Cultivated Area", xlab="Crop Commercialization 3-class")
plot(svytable(~croparea_4ha+noaggross_clas, ghs.svy.shf), main=NA, las=1, col=viridis(3),
  ylab="Cultivated Area", xlab="Non-Farm Income 3-class")
plot(svytable(~croparea_4ha+class9, ghs.svy.shf), main=NA, las=1, col=viridis(9),
  ylab="Cultivated Area", xlab="Typology 9-class")

```


```{r nga-tab1, results="asis"}

tmp <- svymean(~I(interaction(croparea_4ha, cropsales_clas, noaggross_clas)), ghs.svy.farm, na.rm=T)
tmp.cv <- confint(tmp)
tmp <- cbind(as.data.table(tmp*100), as.data.table(tmp.cv*100))
tmp[`2.5 %`<0, `2.5 %` := 0]
tmp <- tmp[, lapply(.SD, function(x) format(x, digits=0, nsmall=1, scientific=F))]
tmp[, confint := paste(`2.5 %`, `97.5 %`, sep=" - ")]

tmp[, `:=`(
  croparea_4ha = factor(rep(levels(ghs$croparea_4ha), 9)),
  cropsales_clas = factor(rep(levels(ghs$cropsales_clas), each=2, 3), levels=c("LoC", "MeC", "HiC")),
  noaggross_clas = factor(rep(levels(ghs$noaggross_clas), each=6, 1), levels=c("LoD", "MeD", "HiD"))
  )]

table_options(HTMLcaption="(#tab:tab1) Est. Proportions of Farm Holdings across Categories (percent)")
html(tabular(Heading()*noaggross_clas*(mean+confint)~Heading()*croparea_4ha*Heading()*cropsales_clas*Heading()*identity,
  data=tmp), rmarkdown=T, 
  caption="(#tab:tab1) Est. Proportions of Farm Holdings across Categories (percent)")

```


> Using the **<= 4 ha** smallholding definition above, we estimate as of 2012:
> 
> * there are **33M** households in Nigeria
> * **58.8%** (**19.2M**) generate a revenue from farming (wages, rents, crops, livestock, and/or fish)
> * **97%** (**16M**) of farm holdings are *small*
> * corresponding to a beneficiary population of **107M**
> * crop commercialization is low at **12.2%** for smallholdings (below 4 ha) and **17.7%** for larger holdings (over 4 ha) on average. Median rates are much smaller however with **0.07%** for smallholdings and **0.30%** for larger holdings.
> * income diversification in farm households is **17%** on average (**18%** for smallholdings and **13%** for larger holdings) (i.e. off-farm income is 17% on average), but medians off-farm incomes are negligible at **0.83%** (**0.63%** for smallholdings and **0.33%** for larger holdings).
> 
> <u>Notes</u>: for LCU conversions **1 int$ = 108.23 Naira** (2015)


```{r nga-fig1b, fig.cap="Prob. Density of Cultivated Area, Crop Commercialization, and Income Diversification"}

par(mfrow=c(1,3), font.main=1)
svyhist(~croparea, ghs.svy.farm, main="Cultivated Area", xlab="croparea (ha)", xlim=c(0,10))
lines(svysmooth(~croparea, ghs.svy.farm, bandwidth=1), lwd=1, col="red")
svyhist(~I(cropsales_sh*100), ghs.svy.farm, main="Crop Commercialization", xlab="cropsales_sh (percent)")
lines(svysmooth(~I(cropsales_sh*100), ghs.svy.farm, bandwidth=1), lwd=1, col="red")
svyhist(~I(noaggross_sh*100), ghs.svy.farm, main="Income Diversification", xlab="noaggross_sh (percent)")
lines(svysmooth(~I(noaggross_sh*100), ghs.svy.farm, bandwidth=1), lwd=1, col="red")

```

```{r nga-fig1c, fig.cap="Crop Commercialization and Income Diversification along Cultivated Area"}

par(mfrow=c(1,2), cex=.9)
svyboxplot(I(cropsales_sh*100)~croparea_4ha, ghs.svy.farm,
  xlab="Cultivated Area", ylab="Crop Commercialization (percent)", legend=2, ylim=c(0,20))
svyboxplot(I(noaggross_sh*100)~croparea_4ha, ghs.svy.farm,
  xlab="Cultivated Area", ylab="Off-Farm Incomen (percent)", legend=2, ylim=c(0,20))

```


```{r nga-tab2, results="asis"}

# Estimated quintiles of `cropsales_sh` for smallholders
tmp <- svyCrossTab(list(~I(cropsales_sh*100), ~I(cropsales/1000)), ~croparea_4ha,
  ghs.svy.farm, quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:tab2a) Crop Commercialization for Farms below/above 4 ha (percent, '000 LCU)")
html(tabular(
  Variable*(Mean+Q25+Q50+Q75)~Heading()*By*Heading()*Stat*Heading()*Format(big.mark=",", nsmall=2, digits=0)*identity, 
  data=tmp))

# Estimated quintiles of `noaggross_sh` for smallholders
tmp <- svyCrossTab(list(~I(noaggross_sh*100), ~I(noaggross/1000)), ~croparea_4ha,
  ghs.svy.farm, quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:tab2b) Non-Farm Income for Farms below/above 4 ha (percent, '000 LCU)")
html(tabular(
  Variable*(Mean+Q25+Q50+Q75)~Heading()*By*Heading()*Stat*Heading()*Format(big.mark=",", nsmall=2, digits=0)*identity,
  data=tmp))

```

The 9 categories of small farm holdings are further characterized in the following graphs and summary tables.

```{r nga-fig1e, fig.cap="Crop Commercialization by Age and Gender", dev="png"}

par(mfrow=c(1,2), cex=.9)
svyplot(I(cropsales/1000)~agehead, subset(ghs.svy.farm, femhead==0), 
  style="transparent", basecol=viridis(2)[1], 
  xlab="Age of Hhld Head", ylab="Crop Sales ('000 LCU)")
legend("topright", lty=c(1,1), col=viridis(2)[1], legend=c("Male"), bty="n")
svyplot(I(cropsales/1000)~agehead, subset(ghs.svy.farm, femhead==1), 
  style="transparent", basecol=viridis(2)[2], 
  xlab="Age of Hhld Head", ylab="Crop Sales ('000 LCU)")
legend("topright", lty=c(1,1), col=viridis(2)[2], legend=c("Female"), bty="n")

```

```{r nga-fig1f, eval=F, fig.cap="Crop Commercialization by Age and Zone"}

par(mfrow=c(1,1), cex=.9)
l <- levels(ghs$zone)
plot(c(15,80), c(0,350), type="n", 
  xlab="Age of Hhld Head", ylab="Est. Mean Crop Sales ('000 LCU)")
legend("topright", lty=c(1,1), col=viridis(5), legend=l, bty="n")
for (i in seq_along(l)) lines(
  svysmooth(I(cropsales/1000)~agehead, subset(ghs.svy.farm, zone==levels(ghs$zone)[i]), banwidth=10), col=viridis(5)[i])

```

```{r nga-fig1g, fig.cap="Income Diversification by Age and Gender", dev="png"}

par(mfrow=c(1,2), cex=.9)
svyplot(I(noaggross/1000)~agehead, subset(ghs.svy.farm, femhead==0), 
  style="transparent", basecol=viridis(2)[1], 
  xlab="Age of Hhld Head", ylab="Non-Farm Income ('000 LCU)")
legend("topright", lty=c(1,1), col=viridis(2)[1], legend=c("Male"), bty="n")
svyplot(I(noaggross/1000)~agehead, subset(ghs.svy.farm, femhead==1), 
  style="transparent", basecol=viridis(2)[2], 
  xlab="Age of Hhld Head", ylab="Non-Farm Income ('000 LCU)")
legend("topright", lty=c(1,1), col=viridis(2)[2], legend=c("Female"), bty="n")

```

```{r, eval=F, fig.cap="Income Diversification by Age and Zone"}

l <- levels(ghs$zone)
plot(c(15,100), c(0,800), type="n", 
  xlab="Age of Hhld Head", ylab="Non-Farm Income ('000 LCU)")
legend("topright", lty=c(1,1), col=viridis(5), legend=l, bty="n")
for (i in seq_along(l)) lines(
  svysmooth(I(noaggross/1000)~agehead, subset(ghs.svy.farm, zone==levels(ghs$zone)[i]), banwidth=10), col=viridis(5)[i])

```


```{r nga-fig1h, fig.cap="Crop Commercialization and Income Diversification along Market Access", dev="png"}

par(mfrow=c(1,2), cex=.9)
svyplot(I(cropsales/1000)~dist_market, ghs.svy.farm, 
  style="transparent", basecol=viridis(2)[1], 
  xlab="Distance to Market (km)", ylab="Crop Sales ('000 LCU)")
svyplot(I(noaggross/1000)~dist_market, ghs.svy.farm, 
  style="transparent", basecol=viridis(2)[2], 
  xlab="Distance to Market (km)", ylab="Non-Farm Income ('000 LCU)")

```

```{r nga-fig1i, fig.cap="Crop Commercialization and Income Diversification vs. Distance to Nearest Road", dev="png"}

par(mfrow=c(1,2), cex=.9)
svyplot(I(cropsales/1000)~dist_road2, ghs.svy.farm, 
  style="transparent", basecol=viridis(2)[1], 
  xlab="Distance to Road (km)", ylab="Crop Sales ('000 LCU)")
svyplot(I(noaggross/1000)~dist_road2, ghs.svy.farm, 
  style="transparent", basecol=viridis(2)[2], 
  xlab="Distance to Road (km)", ylab="Non-Farm Income ('000 LCU)")

```

```{r nga-fig3, eval=F, fig.cap="Distribution of Farm Sizes across Categories of Small Farm Holdings (ha)", fig.height=6}

# farmarea
svyCrossPlot(~farmarea, ghs.svy.shf, breaks=20, bandwidth=.5, xlim=c(0,4),
  xlab="Farm Area (ha)")

```

```{r nga-fig4, eval=F, fig.cap="Distribution of Livestock Assets across Categories of Small Farm Holdings (TLU)", fig.height=6}

# TLU_total
svyCrossPlot(~TLU_total, ghs.svy.shf, breaks=20, bandwidth=1, xlim=c(0,30),
  xlab="Livestock Holdings (TLU)")

```

```{r nga-fig5, eval=F, fig.cap="Distribution of Livestock Sales across Categories of Small Farm Holdings ('000 LCU)", fig.height=6}

# lstocksold
svyCrossPlot(~I(lstocksold/1000), ghs.svy.shf, breaks=20, bandwidth=100, xlim=c(0,1500),
  xlab="Livestock Sales ('000 LCU)")

```

```{r nga-fig6, eval=F, fig.cap="Share of Farm Income across Categories of Small Farm Holdings (percent)", fig.height=6}

# aggross_sh
svyCrossPlot(~I(aggross_sh*100), ghs.svy.shf, breaks=20, bandwidth=1,
  xlab="Share of Farm Income (percent)")

```

Note that all tables below omit standard errors for clarity.

```{r nga-tab3, results="asis"}

# Market access
tmp <- svyCrossTab(list(~dist_market, ~dist_road2, ~mkt50k, ~mkt100k),
  ~class9, ghs.geo.svy.shf, quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:tab3a) Market Access for Small Farm Households across Categories (km, travel time in hrs)")
html(tabular(
  Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=1, nsmall=2, big.mark=",", scientific=F),
  data=tmp)[, -seq(2,18,2)])

# Sources of income
tmp <- svyCrossTab(list(~I(cropvalue/1000), ~I(cropsales/1000), ~I(lstocksold/1000),
  ~I(agsales/1000), ~agsalespc, ~I(aggross/1000), 
  ~I(cropsales_sh*100), ~I(aggross_sh*100)),
  ~class9, ghs.svy.shf, quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:tab3b) Sources of Agricultural Income for Small Farm Households across Categories ('000 LCU, percent)")
html(tabular(
  Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, big.mark=",", scientific=F),
  data=tmp)[, -seq(2,18,2)])

# Livestock
tmp <- svyCrossTab(list(~farmarea, ~croparea, ~catt_numberimp, ~pigs_numberimp,
  ~chic_numberimp, ~sheep_numberimp, ~goat_numberimp, ~TLU_total),
  ~class9, ghs.svy.shf, quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:tab3c) Land and Livestock Assets of Small Farm Households across Categories (ha, TLU)")
html(tabular(
  Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=1, big.mark=",", scientific=F),
  data=tmp)[, -seq(2,18,2)])


```

```{r nga-tab4, results="asis"}

# Demographics
tmp <- svyCrossTab(list(~hhsize, ~I(femhead*100), ~agehead, ~I(lithead*100),
  ~eduyears, ~numchildren, ~equiv, ~females, ~males, 
  ~sh_ed_none, ~sh_ed_prim, ~sh_ed_secu),
  ~class9, ghs.svy.shf, quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:tab4) Demographics of Small Farm Households across Categories")
html(tabular(
  Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=1, big.mark=",", scientific=F),
  data=tmp)[, -seq(2,18,2)])

```


```{r nga-tab5, results="asis"}

# Input uses
tmp <- svyCrossTab(list(~parcels, ~I(chempest_farms*100), ~I(chemherb_farms*100),
  ~I(fertorg_farms*100), ~I(fertinorg_farms*100), ~I(seedp_farms*100), ~I(tractorown_farms*100), ~intens_idx2),
  ~class9, ghs.farm.svy.shf, quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:tab5) Parcels and Input Uses for Small Farm Households across Categories (percent)")
html(tabular(
  Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=1, big.mark=",", scientific=F),
  data=tmp)[, -seq(2,18,2)])

```

```{r nga-tab6, results="asis"}

# Poverty
tmp <- svyCrossTab(list(~pcexp_ppp_m, ~foodexp_ppp_m, ~nfoodexp_ppp_m,
  ~I(poor_ppp1*100), ~I(poor_ppp2*100), ~I(povgap_ppp1*100), ~I(povgap_ppp2*100),
  ~I(sevpov_ppp1*100), ~I(sevpov_ppp2*100)),
  ~class9, ghs.ppp.svy.shf, quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:tab6) Poverty Incidence in Small Farm Households across Categories (PPP, percent)")
html(tabular(
  Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, big.mark=",", nsmall=1, scientific=F),
  data=tmp)[, -seq(2,18,2)])

```
