==> R/04-quadrant.Rmd <==
# Quadrant Classification

## Nigeria

[TBD]

For documentation purposes below are code snippets from the 2016 Farmer Segmentation for Nigeria (definitions and characterization of low/high quadrants and livelihood zones).


```{r nga-quad, eval=FALSE}

#####################################################################################
# 2016.05.03 Quadrant Definitions across Livelihood Zones (from Diemuth 2016.04.06)
#####################################################################################
# We need to add a new segment vars to all key micro datasets, starting with GHS.
# Diemuth provided latest quadrant definitions.
#
#  - [country]_agpotallclass_1km.tif – Agricultural potential raster with 1km cell
# resolution and same extent as ethtt_1km100khr. Created from clipping sub-Saharan Africa
# Agricultural Potential 5 minute grid vector layer from FAO & IIASA, 2007. For the
# 2x2 segmented data, 5-6 values were considered high, 1-4 low and -809-0 were
# unsuitable and -997 were water.
# -997 = Water
# -809 = Protected Areas
# -807 = Urban Areas
# -805 = Irrigated Areas
# -801 = Forest
# 0 = Land not suited for agriculture
# 1 = Land very poorly suited for pasture and at best poorly suited for rainfed crops
# 2 = Land poorly suited for pasture and at best poorly suited for rainfed crops
# 3 = Lands suited for pasture and at best poorly suited for rainfed crops
# 4 = Land suited for rainfed crops and pasture possible
# 5 = Land well suited for rainfed crops and pasture possible
# 6 = Prime land suited for rainfed crops and pasture possible
#
# - [country]_tt1km100khr_4hr.tif – Classified travel time raster with a 4 hr threshold,
# HarvestChoice, 2015
# 100 – travel time greater than 4 hours, lo access
# 200 – travel time less than or equal to 4 hours, high access

library(raster)
library(data.table)
library(viridis)

setwd("~/Google Drive/2016-BMGF-Segmentation/NGA")

# Load the 2 rasters and classify LGAs
qd.agpot <- raster("./maps/quadrant/nga_agpotallclass_1km.tif")
qd.tt100k <- raster("./maps/quadrant/nga_tt1km100khr_4hr.tif")
proj4string(qd.agpot)
proj4string(qd.tt100k)
proj4string(geo12.pts)
unique(qd.agpot)
#  [1] -997 -809 -807 -805 -801    0    1    2    3    4    5    6
summary(qd.tt100k)
#         nga_tt1km100khr_4hr
# Min.                    100
# 1st Qu.                 100
# Median                  200
# 3rd Qu.                 200
# Max.                    200
# NA's                      0
# Note that raster tt100k is already classified.

qd.agpot.lbl <- c(
  "water",
  "protected area",
  "urban area",
  "irrigated area",
  "forest",
  "land not suited for agriculture",
  "land very poorly suited for pasture and at best poorly suited for rainfed crops",
  "land poorly suited for pasture and at best poorly suited for rainfed crops",
  "land suited for pasture and at best poorly suited for rainfed crops",
  "land suited for rainfed crops and pasture possible ",
  "land well suited for rainfed crops and pasture possible",
  "prime land suited for rainfed crops and pasture possible")

qd.agpot.lbl <- data.table(ID=1:12, code=unique(qd.agpot), label=qd.agpot.lbl)

# Plot agpot layer
spplot(qd.agpot, at=unique(qd.agpot), col.regions=viridis(12), colorkey=list(at=1:12, labels=as.character(unique(qd.agpot))))


# Intersect with GHS coords
tmp <- extract(qd.agpot, geo12.pts)
tmp <- data.table(hhid=geo12.pts$hhid, qd.agpot=tmp)
tmp[, qd.agpot := factor(qd.agpot, levels=qd.agpot.lbl$varCode, labels=qd.agpot.lbl$varLabel)]
tmp1 <- extract(qd.tt100k, geo12.pts)
tmp[, qd.tt100k := tmp1]
tmp[, qd.tt100k := factor(qd.tt100k, levels=c(100, 200), labels=c("ML", "MH"))]
setkey(tmp, hhid)
setkey(geo12, hhid)
geo12 <- tmp[geo12]
rm(tmp, tmp1)

# Classify hhlds across quadrants
# Rename old segment var to avoid confusion
setnames(geo12, c("qd.agpot", "qd.tt100k"), c("qd_agpot", "qd_tt100k"))
setnames(geo12, "seg_quad", "seg_quad_old")
setnames(ghs, "seg_quad", "seg_quad_old")
setnames(ppp12, "seg_quad", "seg_quad_old")
setnames(inc12, "seg_quad", "seg_quad_old")
setnames(farm, "seg_quad", "seg_quad_old")

geo12[qd_agpot %in% qd.agpot.lbl$varLabel[1:6], qd_agpot_clas := "AU"]
geo12[qd_agpot %in% qd.agpot.lbl$varLabel[7:10], qd_agpot_clas := "AL"]
geo12[qd_agpot %in% qd.agpot.lbl$varLabel[11:12], qd_agpot_clas := "AH"]
summary(factor(geo12$qd_agpot_clas))
#   AH   AL   AU NA's
# 1968 1490 1323  117

summary(factor(geo12$qd_tt100k))
#   ML   MH NA's
#  750 4051   97

# Impute missing hhld coords with LGA means
# Compute LGA means from `g2` map (using dominant class)
geo12[is.na(qd_agpot_clas), .N, by=.(adm1_nm, adm2_nm)]
# => missing records in 58 LGAs

tmp <- geo12[!is.na(qd_agpot_clas), .(
  qd_agpot_clas=modal(qd_agpot_clas, na.rm=T),
  qd_tt100k=modal(qd_tt100k, na.rm=T)
), keyby=.(adm1_nm, adm2_nm)]

setnames(tmp, 3:4, c("qd_agpot_clas_imp", "qd_tt100k_imp"))
setkey(geo12, adm1_nm, adm2_nm)
geo12 <- tmp[geo12]

geo12[is.na(qd_agpot_clas), qd_agpot_clas := qd_agpot_clas_imp]
geo12[is.na(qd_tt100k), qd_tt100k := qd_tt100k_imp]

geo12[is.na(qd_agpot_clas), .N, by=.(adm1_nm, adm2_nm)]
#    adm1_nm adm2_nm  N
# 1:  Rivers   Bonny 10
geo12[is.na(qd_tt100k), .N, by=.(adm1_nm, adm2_nm)]
# Empty data.table (0 rows) of 3 cols: adm1_nm,adm2_nm,N
# => only 10 missing records in Bonny

# What is the dominant agpot in Bonny? Extract all LGAs for re-use.
tmp <- extract(qd.agpot, g2, fun=modal, na.rm=T)
g2$qd_agpot <- tmp
# Note that we already have median 100k travel time across LGAs in `g2$mkt100k`
# though I'm not sure it's the same source as Diemuth/Stan

# load Joe's latest travel times
tt100k <- dt[, .SD, .SDcols=c("CELL5M", "X", "Y", "tt10_100k")]
rm(dt)
tt100k <- SpatialPixelsDataFrame(tt100k[, .(X,Y)], data.frame(tt100k))
proj4string(tt100k) <- CRS("+init=epsg:4326")
tt100k <- raster(tt100k, layer="tt10_100k")
tmp <- extract(tt100k, g2, fun=median, na.rm=T)
g2$qd_tt100k <- tmp
# Export for mapping
tt100k <- crop(tt100k, g2)
tt100k <- mask(tt100k, g2)
writeRaster(tt100k, "./maps/markets/tt100k.tif")
rm(tmp)

# What is Bonny?
g2[g2$adm2_nm=="Bonny", "qd_agpot"]
# => -997, so AU
geo12[is.na(qd_agpot_clas), qd_agpot_clas := "AU"]
summary(factor(geo12$qd_agpot_clas))
#   AH   AL   AU
# 1997 1530 1371

# Make final quadrant classification for GHS
geo12[, seg_quad := paste0(qd_tt100k, qd_agpot_clas)]
geo12[, unique(seg_quad)]
# [1] "MHAU" "MHAL" "MHAH" "MLAH" "MLAL" "MLAU"
geo12[, seg_quad := ordered(seg_quad, levels=c("MLAU", "MLAL", "MLAH", "MHAU", "MHAL", "MHAH"))]

# Merge quadrants into `ghs`, `farm` and `ppp12`
setkey(geo12, hhid)
setkey(ghs, hhid)
setkey(farm, hhid)
setkey(ppp12, hhid)
ghs$seg_quad <- geo12[ghs][, seg_quad]
farm$seg_quad <- geo12[farm][, seg_quad]
ppp12$seg_quad <- geo12[ppp12][, seg_quad]

# Re-run summary tables
# => done

#####################################################################################
# Crop preferences across quadrants/zones - GHS
#####################################################################################

# Note that Cleo generated the hh-level vars and Eduardo estimated state-level vars
ghs.plot <- read.dta("./data/NGA-GHS/NGA_GHS_Indicator_plot_2012.dta")
ghs.plot <- data.table(ghs.plot)

# Merge in segmentation vars
setkey(geo12, hhid)
setkey(ghs.crop, hhid)
setkey(ghs.plot, hhid)
ghs.crop$seg_quad <- geo12[ghs.crop][, seg_quad]
ghs.plot$seg_quad <- geo12[ghs.plot][, seg_quad]
ghs.crop$seg_lvl_6 <- geo12[ghs.crop][, seg_lvl_6]
ghs.plot$seg_lvl_6 <- geo12[ghs.plot][, seg_lvl_6]

# Share of female plot managers
tmp <- ghs.plot[, .N, by=.(hhid, wt_w2v1, manager, state, gender)]

tmp <- ghs.crop[, .(
  num_fem_mgr=sum(num_fem_mgr*wt_wave2, na.rm=T)/sum(num_hh_mgr*wt_wave2, na.rm=T),
  fem_mgr=weighted.mean(num_fem_mgr>0, wt_wave2, na.rm=T),
  fem_earn=weighted.mean(fem_earn, wt_wave2, na.rm=T)
), keyby=seg_lvl_6]
tmp <- ghs.crop[, .(
  num_fem_mgr=sum(num_fem_mgr*wt_wave2, na.rm=T)/sum(num_hh_mgr*wt_wave2, na.rm=T),
  fem_mgr=weighted.mean(num_fem_mgr>0, wt_wave2, na.rm=T),
  fem_earn=weighted.mean(fem_earn, wt_wave2, na.rm=T)
), keyby=seg_quad]
tmp <- ghs.crop[, .(
  num_fem_mgr=sum(num_fem_mgr*wt_wave2, na.rm=T)/sum(num_hh_mgr*wt_wave2, na.rm=T),
  fem_mgr=weighted.mean(num_fem_mgr>0, wt_wave2, na.rm=T),
  fem_earn=weighted.mean(fem_earn, wt_wave2, na.rm=T))]


# Share of female-managed plots
tmp <- ghs.plot[, .(femplotmgr=weighted.mean(gender=="female", wt_w2v1, na.rm=T)),
  keyby=seg_lvl_6]
tmp <- ghs.plot[, .(femplotmgr=weighted.mean(gender=="female", wt_w2v1, na.rm=T)),
  keyby=seg_quad]
tmp <- ghs.plot[, .(femplotmgr=weighted.mean(gender=="female", wt_w2v1, na.rm=T))]


# TODO Area of female-managed plots under each crop
tmp <- ghs.plot[, lapply(.SD, sum, na.rm=T),
  keyby=.(hhid, wt_w2v1, state, gender), .SDcols=96:147]
tmp <- tmp[, lapply(.SD, function(x) sum(x*wt_w2v1, na.rm=T)),
  keyby=.(state, gender), .SDcols=c(2, 5:56)]
tmp[, wt_w2v1 := NULL]
tmp <- melt(tmp, id.vars=1:2)
tmp[, variable := str_replace(variable, "area_", "")]
tmp <- tmp[!is.na(gender)]

# Proportion of female-managed plots by crop
tmp <- melt(ghs.plot,
  id.vars=c("seg_quad", "seg_lvl_6", "gender", "wt_w2v1"),
  measure.vars=11:147)
tmp[value=="yes", value := "1"]
tmp[value=="no", value := "0"]
tmp[, value := as.numeric(value)]
tmp <- tmp[, lapply(.SD, weighted.mean, wt_w2v1), keyby=.(seg_lvl_6, gender, variable)]

# TODO Additional production stats
# TODO Also check on Sara's harmonized farm management stats

```


## Ghana

```{r gha-quad50k, eval=FALSE}

library(raster)
library(data.table)
library(viridis)
library(stringr)
library(foreign)
library(survey)

setwd("~/Projects/2017-agra-aasr")
load("./tmp/2017-agra-aasr_GHA.RData")

# For now we replicate Nigeria's quadrant classification above to Ghana using both 
# time-to-20K-market and time-to-50K-market layers.
# Load all 3 rasters and classify Ghana districts
qd.tt20k <- raster("~/Maps/cell5m/tt10_SSA_tiff/traveltimetomarket_ssa_020k.tif")
qd.tt50k <- raster("~/Maps/cell5m/tt10_SSA_tiff/traveltimetomarket_ssa_050k.tif")
qd.agpot <- raster("~/Maps/GAEZ_res03crav6190lsilrcer/res03_crav6190l_silr_cer.tif")

proj4string(qd.agpot)
# [1] "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
proj4string(qd.tt20k)
# [1] "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
proj4string(g2)
# [1] "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
res(qd.agpot)
# [1] 0.08333333 0.08333333
res(qd.tt20k)
# [1] 0.08333333 0.08333333
levels(qd.agpot) # => not classified

# Crop all layers to Ghana
qd.agpot <- crop(qd.agpot, g2)
qd.tt20k <- crop(qd.tt20k, g2)
qd.tt50k <- crop(qd.tt50k, g2)

# GAEZ color key from the documentation
unique(qd.agpot)
# [1] 1 2 3 4 5 6 7 8 9
qd.agpot.lbl <- c(
  `very high`=1,
  `high`=2,
  `good`=3,
  `medium`=4,
  `moderate`=5,
  `marginal`=6,
  `very marginal`=7,
  `not suitable`=8,
  `water`=9)

pal.agpot <- fread("
0	112	0	Dark green
21	174	17	Lime green
172	225	94	Conifer
254	216	1	Gold
206	157	83	Tussock
205	108	1	Sun
128	128	128	Aluminium
228	228	228	Athens grey
32	32	255	Blue")
pal.agpot <- rgb(pal.agpot[, .(V1, V2, V3)], maxColorValue=255)

pal.et_anom <- fread("
245	145	31	Lightning yellow
246	169	30	Supernova
252	211	112	Salomie
205	204	204	Blue zodiac
214	232	186	Snow flurry
158	203	59	Atlantis
54	170	73	Forest green
  ")
pal.et_anom <- rgb(pal.et_anom[, .(V1, V2, V3)], maxColorValue=255)
pal.et_anom <- colorRampPalette(pal.et_anom)

pal.ndvi_anom <- fread("
120	81	12	Buttered rum
151	112	59	Muddy waters
176	147	107	Muesli
206	194	177	Stark white
222	222	222	Gainsboro
177	201	167	Spring rain
121	172	106	Gossip
76	150	65	Bilbao
31	128	22	Forest green
  ")
pal.ndvi_anom <- rgb(pal.ndvi_anom[, .(V1, V2, V3)], maxColorValue=255)
pal.ndvi_anom <- colorRampPalette(pal.ndvi_anom)

# qd.agpot.lbl <- c(
#   "water",
#   "protected area",
#   "urban area",
#   "irrigated area",
#   "forest",
#   "land not suited for agriculture",
#   "land very poorly suited for pasture and at best poorly suited for rainfed crops",
#   "land poorly suited for pasture and at best poorly suited for rainfed crops",
#   "land suited for pasture and at best poorly suited for rainfed crops",
#   "land suited for rainfed crops and pasture possible ",
#   "land well suited for rainfed crops and pasture possible",
#   "prime land suited for rainfed crops and pasture possible")

qd.agpot.lbl <- data.table(ID=1:9, code=names(qd.agpot.lbl), col=pal.agpot)
colortable(qd.agpot) <- logical(0)
tm_shape(crop(qd.agpot, g2)) + tm_raster(names(qd.agpot), breaks=1:10,
  labels=qd.agpot.lbl$code, palette=qd.agpot.lbl$col) +
  tm_layout(legend.outside=T)
levels(qd.agpot) <- qd.agpot.lbl

summary(qd.tt20k)
# Min.                      0.0486763
# 1st Qu.                   1.1240800
# Median                    1.7141100
# 3rd Qu.                   2.5925300
# Max.                     12.9055996
# NA's                    350.0000000

# Resample time-to-market layers and GAEZ to 1km (from 10km)
qd.agpot <- disaggregate(qd.agpot, fact=10)
qd.tt20k <- disaggregate(qd.tt20k, fact=10, method="bilinear")
qd.tt50k <- disaggregate(qd.tt50k, fact=10, method="bilinear")

# Verify
# Also load earlier 2008 (instead of 2010) TT rasters
qd.tt08 <- fread("~/Maps/cell5m/tt10_SSA_tiff/hcapi-travel_time-ssa.csv")
qd.tt08 <- SpatialPixelsDataFrame(qd.tt08[, .(X,Y)], 
  data.frame(qd.tt08)[, 8:17], proj4string=CRS("+init=epsg:4326"))
qd.tt08 <- brick(qd.tt08)

# Resample
qd.tt08 <- crop(qd.tt08, g2[g2$svyCode=="gha-glss6",])
qd.tt08 <- disaggregate(qd.tt08, fact=10)

spplot(qd.tt08, col.regions=rev(viridis(28)), maxpixels=100000)
spplot(tmp[[1:5]], col.regions=plasma, maxpixels=100000, at=c(seq(-1,24,1), 62),
  xlab="HarvestChoice Travel Times (circa 2007, hrs)",
  par.settings=list(strip.background=list(col="#fafafa")))

# Slow, so save PNG
png("./docs/fig/HC_tt10.png", width=6, height=5, res=220, units="in", family="Roboto Condensed", pointsize=9)
spplot(mask(qd.tt08[[6:10]], g2[g2$svyCode=="gha-glss6",]),
  col.regions=plasma, maxpixels=300000, at=c(seq(-1,24,1), 62),
  xlab=list("IFPRI/HarvestChoice Travel Times (circa 2010, hrs)", cex=.8),
  strip=strip.custom(bg="#fafafa"), par.strip.text=list(cex=.7))
dev.off()

# Keep 2 versions (10 layers) instead.
rm(qd.tt20k, qd.tt50k)

#####################################################################################
# Compare with SDI version received on 2017.06.12
g2.qd <- fread("./out/SDI/gha-L2_stats.csv")
setkey(g2.qd, svyCode,svyL1Cd, svyL2Cd)
setkey(g2.dt, svyCode,svyL1Cd, svyL2Cd)
g2.dt[, seg_quad := NULL]
g2.dt[g2.qd, seg_quad := statsmedia]
unique(g2.dt$seg_quad)
# [1] NA  3  4  0  1  2
# lo-lo to hi-hi, where the first is ag potential and the second is market access
g2.dt[, seg_quad := factor(seg_quad, levels=1:4, 
  labels=c(
    "lo suit.-lo mkt.", 
    "lo suit.-hi mkt.", 
    "hi suit.-lo mkt.", 
    "hi suit.-hi mkt."))]

g2 <- SpatialPolygonsDataFrame(g2, data.frame(g2.dt), match.ID="rn")
spplot(g2[g2$svyCode=="gha-glss6", "seg_quad"], col.regions=pal.agpot[c(5,6,3,2)])
# => TODO should not be any district with NA value
rm(g2.qd)

# Load SSA-wide CELL5M classified shapefile (received from Todd on 2017.06.20)
qd.quad.dt <- fread("~/Maps/GAEZ_res03crav6190lsilrcer/Ag_Suitability_SSA.csv")
names(qd.quad.dt)
# We only need a few fields
qd.quad.dt <- qd.quad.dt[, .(CELL5M=Cell5M_int, agpot, mktacc, seg=quadrant)]

# Verify classification
qd.quad.dt[, .(agpot=unique(agpot)), keyby=seg]
#    seg agpot
# 1:  NA      
# 2:   1    lo
# 3:   2    lo
# 4:   3    hi
# 5:   4    hi

# Merge in X,Y
tmp <- fread("./data/hcapi-X--ssa.csv")
names(tmp)
setkey(tmp, CELL5M)
setkey(qd.quad.dt, CELL5M)
qd.quad.dt <- qd.quad.dt[tmp]

# Convert to raster and map
qd.quad.dt[, .N, by=agpot]
#    agpot      N
# 1:    NA   5099
# 2:        54560
# 3:    lo 118437
# 4:    hi 125357
qd.quad.dt[agpot=="", agpot := "unsuitable"]
qd.quad.dt[, agpot := factor(agpot, levels=c("unsuitable", "lo", "hi"))]
qd.suit <- SpatialPixelsDataFrame(qd.quad.dt[, .(X, Y)],
  data.frame(suit=qd.quad.dt$agpot), proj4string=CRS("+init=epsg:4326"))
qd.suit <- raster(qd.suit)
levels(qd.suit)
# [[1]]
#   ID     levels
# 1  1 unsuitable
# 2  2         lo
# 3  3         hi
spplot(crop(qd.suit, g2), col.regions=pal.ndvi_anom(3)[c(3,1,2)])

# Compare this classified to raw GAEZ


# Resample to 1km (note that resampling loses RAT)
qd.suit <- disaggregate(qd.suit, fact=10)
res(qd.suit)
table(values(qd.suit))
#       1        2        3 
# 5456000 11843700 12535700 
levels(qd.suit) <- data.frame(ID=1:3, cat=c("unsuitable", "low", "high"))

# Save SSA-wide suitability raster for re-use
writeRaster(qd.suit, "~/Maps/BMGF/agpot-1km-lohi_SSA.tif", datatype="INT2S", 
  options=c("COMPRESS=NONE", "TFW=YES"), prj=T, overwrite=T)

qd.suit <- crop(qd.suit, g2)

##################################################################################### 
# Compare with hi-res land cover (ESA/CCI, 2015)
qd.lc <- raster("~/Maps/ESA/ESACCI-LC-L4-LCCS-Map-300m-P1Y-2015-v2.0.7.tif")
res(qd.lc)
# [1] 0.002777778 0.002777778
qd.lc <- crop(qd.lc, g2)
qd.lc[qd.lc==0] <- NA

qd.lc.lbl <- fread("~/Maps/ESA/ESACCI-LC-Legend.csv")
setnames(qd.lc.lbl, 1:2, c("code", "class"))
qd.lc.lbl[, col := rgb(cbind(R,G,B), maxColorValue=255)]
qd.lc.lbl[, cat := str_replace_all(str_replace_all(abbreviate(toupper(class), 5), fixed("("), ""), fixed("/"), "")]
setkey(qd.lc.lbl, code)

png("./docs/fig/gha-lc-maps.png")
spplot(qd.lc, col.regions=qd.lc.lbl[J(unique(qd.lc)), col], at=c(-1, unique(qd.lc)), 
  colorkey=list(at=1:23, labels=list(labels=qd.lc.lbl[J(unique(qd.lc)), cat], at=1:22+.5)))
dev.off()

# Compare with another hi-res land cover (USGS/EROS, 2013)
qd.usgs <- raster("~/Maps/USGS/west_africa_land-use_land-cover_2013_2km/swa_2013lulc_2km.tif")
proj4string(qd.usgs)
# [1] "+proj=laea +lat_0=11 +lon_0=3 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
res(qd.usgs)
# [1] 2000 2000

g2 <- spTransform(g2, proj4string(qd.usgs))
qd.usgs <- crop(qd.usgs, g2)
g2 <- spTransform(g2, proj4string(qd.tt20k))
unique(qd.usgs)
#  [1]  1  2  3  4  6  7  8  9 10 11 12 13 14 15 21 22 23 24 25 27 31 78

qd.usgs.lbl <- data.table(levels(qd.usgs)[[1]])
qd.usgs.lbl[, col := rgb(cbind(red,green,blue), maxColorValue=255)]
setkey(qd.usgs.lbl, value)
levels(qd.usgs) <- qd.usgs.lbl[, .(ID=value, code, class=class_name, col)]
spplot(qd.usgs, "code", col.regions=setorder(qd.usgs.lbl, code)[, col])


##################################################################################### 
# TODO Compare with hi-res population density (Worldpop)
qd.pop <- raster("~/Maps/Worldpop/AFR_PPP_2015_adj_v2.tif")
res(qd.pop)
qd.pop <- crop(qd.pop, g2)
summary(qd.pop)


#####################################################################################
# Summarize all rasters across GLSS districts (3 waves)
proj4string(g2)
# [1] "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
proj4string(qd.suit)
# [1] "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

# We need the following summary stats:
# - median tt20K travel time (qd.tt20k)
# - median tt50K travel time (qd.tt50k)
# - dominant LC class (qd.lc)
# - dominant GAEZ class (qd.agpot)
# - dominant hi/lo suitability class (qd.suit)
# All layers have been cropped to `g2` and resampled to 1km
tmp1 <- extract(qd.tt20k, g2, median, na.rm=T, small=T)
tmp2 <- extract(qd.tt50k, g2, median, na.rm=T, small=T)
tmp3 <- extract(qd.lc, g2, modal, na.rm=T, small=T)
tmp4 <- extract(qd.agpot, g2, modal, na.rm=T, small=T, factors=T, df=T)
tmp5 <- extract(qd.suit, g2, modal, na.rm=T, small=T, factors=T, df=T)
tmp6 <- extract(qd.tt20k, g2, mean, na.rm=T, small=T)
tmp7 <- extract(qd.tt50k, g2, mean, na.rm=T, small=T)

table(tmp3)
#  10  30  40  50  62 120 170 190 210 
# 204   2   2   2  39  21   1  10   4 
table(tmp4)
#  3   4   5   6   9 
# 32  67 163  17   6 
table(tmp5)
#  1   3 
# 31 254 
# => no low suitability district (hum)

# Combine all extracts
g2.dt <- data.table(g2@data, 
  tt20k_med=tmp1[,1], tt50k_med=tmp2[,1], tt20k_mean=tmp6[,1], tt50k_mean=tmp7[,1],
  lc=tmp3[,1], agpot=tmp4[,2], suit=tmp5[,2])

# Convert codes to factors
g2.dt[, lc := factor(lc, levels=qd.lc.lbl$code, labels=qd.lc.lbl$cat)]
summary(g2.dt$lc)
summary(g2.dt$tt20k_mean)
#   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.1076  0.9807  1.4460  1.5450  2.0300  4.7310 
summary(g2.dt$tt50k_mean)
#   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.1082  1.2240  1.8440  1.9390  2.4280  5.7260 
summary(g2.dt$tt20k_med)
#  Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.104   0.879   1.375   1.445   1.824   5.014 
# => use tt50k for now
summary(g2.dt$suit)
#  high        low unsuitable 
#   254          0         31 

# Make lo/hi quadrants (use 4hr cutoff)
levels(g2.dt$suit) <- c("agpot-hi", "agpot-lo", "agpot-no")
g2.dt[, seg_mkt := factor(tt50k_mean <= 4, levels=c(F,T), labels=c("mkt-lo", "mkt-hi"))]
g2.dt[, seg_quad_6 := paste(suit, seg_mkt, sep=" / ")]
g2.dt[, seg_quad_6 := factor(seg_quad_6, levels=c(
  "agpot-no / mkt-lo", "agpot-no / mkt-hi",
  "agpot-lo / mkt-lo", "agpot-lo / mkt-hi",
  "agpot-hi / mkt-lo", "agpot-hi / mkt-hi"))]
g2.dt[, .N, by=seg_quad_6]
#                quad   N
# 1: agpot-hi / mkt-hi 242
# 2: agpot-no / mkt-hi  27
# 3: agpot-hi / mkt-lo  12
# 4: agpot-no / mkt-lo   4

# Reclass all `agpot-no` districts to `agpot-lo`
g2.dt[, seg_quad_4 := seg_quad_6]
levels(g2.dt$seg_quad_4) <- list(
  `agpot-lo / mkt-lo` = c("agpot-no / mkt-lo", "agpot-lo / mkt-lo"), 
  `agpot-lo / mkt-hi` = c("agpot-no / mkt-hi", "agpot-lo / mkt-hi"),
  `agpot-hi / mkt-lo`="agpot-hi / mkt-lo", 
  `agpot-hi / mkt-hi`="agpot-hi / mkt-hi")
g2.dt[, .N, by=seg_quad_4]
#           seg_quad_4   N
# 1: agpot-hi / mkt-hi 242
# 2: agpot-lo / mkt-hi  27
# 3: agpot-hi / mkt-lo  12
# 4: agpot-lo / mkt-lo   4

# Map it
g2 <- SpatialPolygonsDataFrame(g2, data.frame(g2.dt), match.ID="rn")
spplot(g2[g2$svyCode=="gha-glss6",], "seg_quad_4", col="grey20", lwd=.2, col.regions=viridis(4))

# Save attributes for re-use
attr(g2.dt, "var.labels") <- c("ISO3 country code", "survey code", "region code (used in survey)",
  "region", "district code (used in survey)", "district", "district label",
  "area (sq. km.)", "longitude of centroid", "latitude of centroid",
  "shape ID", "agro-ecological zone (reported in survey)", 
  "travel time to 20K market (hrs, median)", "travel time to 50K market (hrs, median)",
  "travel time to 20K market (hrs, mean)", "travel time to 50K market (hrs, mean)",
  "ESA/CCI land cover code", "FAO/GAEZ suitability code", "BMGF/SDI ag. potential class",
  "low/high market access", "quadrant class (6-class)", "quadrant class (4-class)")
write.dta(g2.dt, "./tmp/gha-glss_L2_att.dta", version=12L)

# Merge district classification into hhld records
setkey(g2.dt, svyCode, svyL1Nm, svyL2Cd)
setkey(gha, svyCode, region, svyL2Cd)
gha[g2.dt, `:=`(
  tt20k_med=i.tt20k_med,
  tt50k_med=i.tt50k_med,
  tt20k_mean=i.tt20k_mean,
  tt50k_mean=i.tt50k_mean,
  lc_esa=i.lc,
  suit_gaez=i.agpot,
  suit_bmgf=i.suit,
  seg_mkt=i.seg_mkt,
  seg_quad_6=i.seg_quad_6,
  seg_quad_4=i.seg_quad_4)]

# Verify
gha[svyCode=="gha-glss6", summary(seg_quad_6)]
gha[, .N, by=.(svyCode, seg_quad_4)]
#      svyCode        seg_quad_4     N
# 1: gha-glss6 agpot-hi / mkt-hi 12905
# 2: gha-glss6 agpot-lo / mkt-hi  3183
# 3: gha-glss6 agpot-hi / mkt-lo   508
# 4: gha-glss6 agpot-lo / mkt-lo   176
# 5: gha-glss5 agpot-hi / mkt-hi  6246
# 6: gha-glss5 agpot-lo / mkt-hi  2066
# 7: gha-glss5 agpot-hi / mkt-lo   300
# 8: gha-glss5 agpot-lo / mkt-lo    75
# 9: gha-glss4                NA  5998

# Update all survey designs
gha.svy <- list(
  `gha6` = svydesign(~clust+hhid, strata=~region+rural, w=~weight, nest=T,
    data=gha[survey=="2012/13"]),
  `gha5` =  svydesign(~clust+hhid, strata=~region+rural, w=~weight, nest=T,
    data=gha[survey=="2005/06"]),  
  `gha4` = svydesign(~clust+hhid, strata=~rural, w=~weight, nest=T,
    data=gha[survey=="1998/99"])
)

# Subset survey sample to farms
gha.svy.farm <- lapply(gha.svy, subset, !is.na(croparea_clas))

# Subset survey sample to only SHF (note that they exclude farms with no `croparea`)
gha.svy.shf <- lapply(gha.svy, subset, croparea_4ha=="<= 4 ha")

```

```{r gha-quad100k, eval=FALSE}

# Use GLUES suitability classification and `tt10_100k` (instead of `tt10_50k`)

# Market access 100K
# Resample time-to-market layers to 1km (from 10km)
res(qd.tt08)
# [1] 0.008333333 0.008333333 => already at 1km

# Summarize across districts
tmp1 <- extract(crop(qd.tt08[["tt10_100k"]], g2), g2, mean, na.rm=T, small=T)
tmp2 <- extract(crop(qd.tt08[["tt10_100k"]], g2), g2, median, na.rm=T, small=T)

# Sumarize GLUES suitability across districts
glues <- raster("~/Maps/GLUES/overall_cropsuit_i_1981-2010.tif")
crs(glues)
NAvalue(glues)
minValue(glues)
maxValue(glues)
tmp3 <- extract(crop(glues, g2), g2, mean, na.rm=T, small=T)
tmp4 <- extract(crop(glues, g2), g2, median, na.rm=T, small=T)

# Combine all extracts
g2.dt <- data.table(g2@data,
  tt100k_mean=tmp1[,1], tt100k_med=tmp2[,1]
  suit_glues_mean=tmp3[,1], suit_glues_med=tmp4[,1])

# Convert extracts to factors
summary(g2.dt$tt100k_mean)
#  Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.355   3.708   5.136   6.192   7.354  42.640 
summary(g2.dt$tt100k_med)
#  Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.200   2.900   4.600   5.564   6.500  45.000
summary(g2.dt$tt500k_mean)
#   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.3997  4.2640  6.4290  7.0550  8.6320 42.7300 
summary(g2.dt$tt500k_med)
#  Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.200   3.600   5.900   6.437   7.700  45.200 
summary(g2.dt$suit_glues_mean)
#  Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 14.07   36.68   47.55   48.64   61.24   80.97 
summary(g2.dt$suit_glues_med)
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.00   35.00   50.00   49.43   66.00   89.00

# Make lo/hi quadrants (use 4hr cutoff, and 50% suitability)
g2.dt[, seg_mkt100k := factor(tt100k_mean <= 4, levels=c(F,T), labels=c("mkt-lo", "mkt-hi"))]
g2.dt[, seg_suit_glues := factor(suit_glues_med <= 50, levels=c(T,F), labels=c("agpot-lo", "agpot-hi"))]
g2.dt[, seg_quad_glues_4 := paste(seg_suit_glues, seg_mkt100k, sep=" / ")]
g2.dt[, seg_quad_glues_4 := factor(seg_quad_glues_4, levels=c(
  "agpot-lo / mkt-lo", "agpot-lo / mkt-hi",
  "agpot-hi / mkt-lo", "agpot-hi / mkt-hi"))]
g2.dt[, .N, by=seg_quad_glues_4]
#     seg_quad_glues_4   N
# 1: agpot-lo / mkt-lo  88
# 2: agpot-hi / mkt-lo 120
# 3: agpot-lo / mkt-hi  64
# 4: agpot-hi / mkt-hi  13

# Map it
g2 <- SpatialPolygonsDataFrame(g2, data.frame(g2.dt), match.ID="rn")
spplot(g2[g2$svyCode=="gha-glss6",], "seg_quad_glues_4", col="grey20", lwd=.2, col.regions=viridis(4))

# Merge district classification into hhld records
setkey(g2.dt, svyCode, svyL1Nm, svyL2Cd)
setkey(gha, svyCode, region, svyL2Cd)
gha[g2.dt, `:=`(
  tt500k_mean = i.tt500k_mean,
  tt500k_med = i.tt500k_med,
  tt100k_mean = i.tt100k_mean,
  tt100k_med = i.tt100k_med,  
  suit_glues_mean = i.suit_glues_mean,
  suit_glues_md = i.suit_glues_med,
  seg_mkt500k = i.seg_mkt500k,
  seg_suit_glues = i.seg_suit_glues,
  seg_quad_glues_4 = i.seg_quad_glues_4)]

# Verify
gha[svyCode=="gha-glss6", summary(seg_quad_glues_4)]
# agpot-lo / mkt-lo agpot-lo / mkt-hi agpot-hi / mkt-lo agpot-hi / mkt-hi 
#              4185              4892              6392              1303 
gha[, .N, by=.(svyCode, seg_quad_glues_4)]
#      svyCode  seg_quad_glues_4    N
# 1: gha-glss6 agpot-lo / mkt-lo 4185
# 2: gha-glss6 agpot-lo / mkt-hi 4892
# 3: gha-glss6 agpot-hi / mkt-lo 6392
# 4: gha-glss6 agpot-hi / mkt-hi 1303
# 5: gha-glss5 agpot-lo / mkt-lo 2441
# 6: gha-glss5 agpot-lo / mkt-hi 2530
# 7: gha-glss5 agpot-hi / mkt-lo 3057
# 8: gha-glss5 agpot-hi / mkt-hi  659
# 9: gha-glss4                NA 5998

###############################################
# Also update/correct the initial `seq_quad_4` using `seg_mkt100k` instead of `seg_mkt100k`
# 1) In the district-level table
g2.dt <- data.table(g2@data)
# Make lo/hi quadrants (use 4hr cutoff, and 50% suitability)
g2.dt[, seg_quad_6 := paste(suit, seg_mkt100k, sep=" / ")]
g2.dt[, seg_quad_6 := factor(seg_quad_6, levels=c(
  "agpot-no / mkt-lo", "agpot-no / mkt-hi",
  "agpot-lo / mkt-lo", "agpot-lo / mkt-hi",
  "agpot-hi / mkt-lo", "agpot-hi / mkt-hi"))]

# Reclass all `agpot-no` districts to `agpot-lo`
g2.dt[, seg_quad_4 := seg_quad_6]
levels(g2.dt$seg_quad_4) <- list(
  `agpot-lo / mkt-lo` = c("agpot-no / mkt-lo", "agpot-lo / mkt-lo"), 
  `agpot-lo / mkt-hi` = c("agpot-no / mkt-hi", "agpot-lo / mkt-hi"),
  `agpot-hi / mkt-lo`="agpot-hi / mkt-lo", 
  `agpot-hi / mkt-hi`="agpot-hi / mkt-hi")
g2.dt[, .N, by=.(svyCode, seg_quad_4)]
#      svyCode        seg_quad_4   N
# 1: gha-glss6 agpot-hi / mkt-lo 117
# 2: gha-glss6 agpot-hi / mkt-hi  34
# 3: gha-glss6 agpot-lo / mkt-hi  14
# 4: gha-glss6 agpot-lo / mkt-lo   5
# 5: gha-glss5 agpot-hi / mkt-hi  21
# 6: gha-glss5 agpot-lo / mkt-hi   8
# 7: gha-glss5 agpot-hi / mkt-lo  82
# 8: gha-glss5 agpot-lo / mkt-lo   4

# 2) In the hhld-level table
summary(gha$suit_bmgf)
# agpot-hi agpot-lo agpot-no     NA's 
#    19959        0     5500     5998 
gha[, seg_mkt100k := factor(tt100k_mean <= 4, levels=c(F,T), labels=c("mkt-lo", "mkt-hi"))]
gha[, seg_quad_6 := paste(suit_bmgf, seg_mkt100k, sep=" / ")]
gha[, seg_quad_6 := factor(seg_quad_6, levels=c(
  "agpot-no / mkt-lo", "agpot-no / mkt-hi",
  "agpot-lo / mkt-lo", "agpot-lo / mkt-hi",
  "agpot-hi / mkt-lo", "agpot-hi / mkt-hi"))]

# Reclass all `agpot-no` districts to `agpot-lo`
gha[, seg_quad_4 := seg_quad_6]
levels(gha$seg_quad_4) <- list(
  `agpot-lo / mkt-lo` = c("agpot-no / mkt-lo", "agpot-lo / mkt-lo"), 
  `agpot-lo / mkt-hi` = c("agpot-no / mkt-hi", "agpot-lo / mkt-hi"),
  `agpot-hi / mkt-lo`="agpot-hi / mkt-lo", 
  `agpot-hi / mkt-hi`="agpot-hi / mkt-hi")
gha[, .N, by=seg_quad_4]
#           seg_quad_4     N
# 1: agpot-hi / mkt-lo 15513
# 2: agpot-hi / mkt-hi  4446
# 3: agpot-lo / mkt-hi  4938
# 4: agpot-lo / mkt-lo   562
# 5:                NA  5998

# Map it
g2 <- SpatialPolygonsDataFrame(g2, data.frame(g2.dt), match.ID="rn")
spplot(g2[g2$svyCode=="gha-glss6",], "seg_quad_4", col="grey20", lwd=.2, col.regions=viridis(4))

###############################################
# Another version using GLUES suitability masking out forest/protected/urban areas
# Replace original GLUES grid with modified version
glues <- raster("./out/MB/2017-glues-suit_gha.tif")
# This quadrant grid was generated on AWS UbuntuGIS at 1km resolution
quad.gha <- raster("./out/MB/2017-quad_gha.tif")
crs(quad.gha)
NAvalue(quad.gha)
minValue(quad.gha)
maxValue(quad.gha)
quad.gha <- ratify(quad.gha)
levels(quad.gha)[[1]]$code = c(0,1,10,11)
levels(quad.gha)[[1]]$label = c(
  "agpot-lo / mkt-lo", "agpot-lo / mkt-hi", 
  "agpot-hi / mkt-lo", "agpot-hi / mkt-hi")
spplot(quad.gha, "code")
tmp <- extract(quad.gha, g2, modal, na.rm=T, small=T, factors=T, df=T)
summary(tmp)

# Update district-level vars
g2.dt <- data.table(g2@data)[, 
  seg_quad_glues_4 := factor(tmp[, "label"], levels=levels(quad.gha)[[1]]$label)]
g2.dt[, seg_suit_glues := seg_quad_glues_4]
levels(g2.dt$seg_suit_glues) <- c("agpot-lo", "agpot-lo", "agpot-hi", "agpot-hi")

# Map it
g2 <- SpatialPolygonsDataFrame(g2, data.frame(g2.dt), match.ID="rn")
spplot(g2[g2$svyCode=="gha-glss6",], "seg_quad_glues_4", col="grey20", lwd=.2, col.regions=viridis(4))

# Merge district classification back into hhld records
setkey(g2.dt, svyCode, svyL1Nm, svyL2Cd)
setkey(gha, svyCode, region, svyL2Cd)
gha[g2.dt, `:=`(
  seg_quad_glues_4 = i.seg_quad_glues_4, 
  seg_suit_glues = i.seg_suit_glues)]

# Save map attributes for re-use
attr(g2.dt, "var.labels") <- c("ISO3 country code", "survey code", "region code (used in survey)",
  "region", "district code (used in survey)", "district", "district label",
  "area (sq. km.)", "longitude of centroid", "latitude of centroid",
  "shape ID", "agro-ecological zone (reported in survey)", 
  "travel time to 20K market (hrs, median)", "travel time to 50K market (hrs, median)",
  "travel time to 20K market (hrs, mean)", "travel time to 50K market (hrs, mean)",
  "ESA/CCI land cover code", "FAO/GAEZ suitability code", "BMGF/SDI ag. potential class",
  "low/high market access (50K)", "quadrant class (6-class)", "quadrant class (4-class)",
  "travel time to 500K market (hrs, mean)", "travel time to 500K market (hrs, median)",
  "GLUES suitability (percent, mean)", "GLUES suitability (percent, median)", 
  "low/high market access (500K)", "low/high suitability (GLUES)", 
  "quadrant class (4-class, GLUES/500K)",
  "travel time to 100K market (hrs, mean)", "travel time to 100K market (hrs, median)",
  "low/high market access (100K)")
write.dta(g2.dt, "./tmp/gha-glss_L2_att.dta", version=12L)

# Update all survey designs
gha.svy <- list(
  `gha6` = svydesign(~clust+hhid, strata=~region+rural, w=~weight, nest=T,
    data=gha[survey=="2012/13"]),
  `gha5` =  svydesign(~clust+hhid, strata=~region+rural, w=~weight, nest=T,
    data=gha[survey=="2005/06"]),  
  `gha4` = svydesign(~clust+hhid, strata=~rural, w=~weight, nest=T,
    data=gha[survey=="1998/99"])
)

# Subset survey sample to farms
gha.svy.farm <- lapply(gha.svy, subset, !is.na(croparea_clas))

# Subset survey sample to only SHF (note that they exclude farms with no `croparea`)
gha.svy.shf <- lapply(gha.svy, subset, croparea_4ha=="<= 4 ha")

```

```{r gha-ndvi, eval=FALSE}

# 2017.07.06 Received CELL5M MODIS NDVI and cropland from Jawoo
ndvi <- fread("~/Maps/cell5m/ndvi/cell5m_ndvi_gnet_descaled.csv")
ndvi[, c(2:3, 5:41) := lapply(.SD, as.numeric), .SDcols=c(2:3, 5:41)]
ndvi[, CELL5M := as.integer(CELL5M)]
ndvi[, .N, by=ISO3]
#    ISO3     N
# 1:  ETH 13295
# 2:  NGA 10724
# 3:  GHA  2751
# 4:  TZA 10405
# Get complete CELL5M grid from `qd.quad.dt`
setkey(qd.quad.dt, CELL5M)
setkey(ndvi, CELL5M)
ndvi <- ndvi[qd.quad.dt, `:=`(CELL5M=i.CELL5M, X=i.X, Y=i.Y)]
ndvi[, NDVI := rowMeans(.SD, na.rm=T), .SDcols=6:41]
setorder(ndvi, X, Y)
tmp <- SpatialPixelsDataFrame(ndvi[, .(X,Y)], ndvi[, .SD, .SDcols=5:42],
  proj4string=CRS("+init=epsg:4326"))
tmp <- brick(tmp)

png("./out/MB/MODIS_NDVI_2001-2013.png", width=7, height=3.6, units="in", res=220, pointsize=8, family="Roboto Condensed")
spplot(tmp[["NDVI"]], col.regions=pal.ndvi(255),
  xlab="Long-term NDVI 2001-2013 (mean of dekadal means, 2001-2013). Source: MODIS.")
dev.off()

png("./out/MB/WorldGrids_CRP_ENS.png", width=7, height=3.6, units="in", res=220, pointsize=8, family="Roboto Condensed")
spplot(tmp[["CRP_ENS"]], col.regions=pal.ndvi_anom(255),
  xlab="Cropland (ensemble probability). Source: WorldGrids.")
dev.off()

```


```{r gha-9class, eval=FALSE}

# Peter chose to re-combine the 9 classes developed in Section 3 into 5 categories
# with:
# - 2-class low/high income diversification <= 1/3, > 1/3
# - 3-class sales share Low (≤ 5%) Medium (5-50%) High (>50%)
# We need to create this scheme here.
# - transitioning farms
# - subsistence
# - pre-commercial
# - specialized
# - diversified commercial

##########################################
# Income diversification dummy `noaggross_2clas` (2-class)
gha[, naggross_2clas := NULL]
gha[, naggross_2clas := cut(naggross_sh, c(-1,1/3,1.1), labels=c("LoD", "HiD"), ordered=T, right=F)]
summary(gha$naggross_2clas)
#   LoD  HiD   NA's 
# 10549 18202  2706 
#   LoD   MeD   HiD  NA's 
# 10549  1934 16268  2706 

##########################################
# Classification dummy `class6` (6-class)
class6 <- c("LoCLoD", "LoCHiD",  "MeCLoD", "MeCHiD", "HiCLoD", "HiCHiD")
gha[, class6 := NULL]
gha[, class6 := factor(paste0(cropsales_clas, naggross_2clas), levels=class6)]
summary(gha$class6)
# LoCLoD LoCHiD MeCLoD MeCHiD HiCLoD HiCHiD   NA's 
#   2400  13954   4387   2065   3762   2183   2706

##########################################
# Classification dummy `class5` (5-class)
gha[, class5 := class6]
levels(gha$class5) <- c("subs.", "trans.", "pre-comm.",
  "trans.", "specd. comm.", "divf. comm.")
summary(gha$class5)
# subs.       trans.    pre-comm. specd. comm.  divf. comm.         NA's 
# 2400        16019         4387         3762         2183         2706 

# Update survey designs (code above)

```

```{r save-gha-quad, eval=FALSE}

# Save hhld-level records
save(gha, file="./tmp/2017-agra-aasr_GHA_merged.RData")

# Save workspace
rm(tmp, tmp1, tmp2, tmp3, tmp4, tmp5, tmp6, tmp7, tmp.cv, x, i)
save.image("./tmp/2017-agra-aasr_GHA.RData")

```

Using a combination of biophysical and infrastructure geospatial layers available over sub-Saharan African we identify zones of low/high agricultural potential and low/high market access in Ghana. This basket classification is useful to further characterize small farm holders across all 4 quadrants, and for targeting interventions and investments.

The input layers used for Ghana are as follows.

### Agricultural Potential

This measure was derived from [FAO/IIASA/GAEZ Crop Suitability Index, 2007](http://www.fao.org/nr/gaez/about-data-portal/agricultural-suitability-and-potential-yields/en/) (cereals, rainfed, low-input). Soil suitability classifications are based on knowledge of crop requirements, of prevailing soil conditions, and of applied soil management. In other words, soil suitability procedures quantify to what extent soil conditions match crop requirements under defined input and management circumstances. 

Low and high potential areas are further classified as:

- 1-4 values are considered **low suitability**: 
    * `1` Land very poorly suited for pasture and at best poorly suited for rainfed crops  
    * `2` Land poorly suited for pasture and at best poorly suited for rainfed crops  
    * `3` Lands suited for pasture and at best poorly suited for rainfed crops  
    * `4` Land suited for rainfed crops and pasture possible  
- 5-6 values are considered **high suitability**:  
    * `5` Land well suited for rainfed crops and pasture possible  
    * `6` Prime land suited for rainfed crops and pasture possible  
- 809-0 is **unsuitable**:  
    * `-809` Protected area  
    * `-807` Urban area  
    * `-805` Irrigated area  
    * `-801` Forest  
    * `0` Land not suited for agriculture  

IIASA/GAEZ crop suitability for Ghana is shown below along with two more recent and higher-resolution remote-sensed land cover classification from [ESA/CCI](http://maps.elie.ucl.ac.be/CCI/viewer/) and from [USGS/EROS](https://eros.usgs.gov/westafrica/data-downloads). The fourth map shows areas of low and high agricultural potential based on additional work conducted by BMGF/SDI (this layer was resampled from 10km to 1km resolution and then summarized across districts).

```{r, fig.cap="Sources of hi-res Land Cover Maps for Ghana"}

include_graphics("./fig/gha-lc-maps.png")

```

```{r, fig.cap="Sources of hi-res Climatic Constraints for Ghana (e.g. FEWSNET/MODISC6)"}

include_graphics("./fig/gha-ndvi-eta_2017.03.d1.png")

```
    
```{r gha-suitmap, fig.cap="Mapping Ghana's Agricultural Potential", dev="png"}

print(spplot(crop(qd.agpot, g2), at=0:9, col.regions=qd.agpot.lbl$col, 
  colorkey=list(labels=list(labels=qd.agpot.lbl$code, at=0:9+.5)),
  xlab=list("Crop Suitability for Low-Input Rainfed Cereals\n(FAO/IIASA/GAEZ 2007)", cex=.8)), 
  split=c(1,1,2,1), more=T)

print(spplot(crop(qd.suit, g2), col.regions=c(qd.agpot.lbl[c(2,6), col], "grey"), 
  xlab=list("Low/High Agricultural Potential\n(BMGF/SDI 2016)", cex=.8)),
  split=c(2,1,2,1), more=F)

```

```{r}

kable(qd.usgs.lbl[, .(Code=value, Label=code, Class=class_name)],
  caption="USGS/EROS Land Cover Classification")

kable(qd.lc.lbl[, .(Code=code, Label=cat, Class=class)],
  caption="ESA/CCI Land Cover Classification")

```
    
### Market Access

Low and high market access areas are defined using a **4 hour** travel time threshold. The source is [IFPRI/HarvestChoice Travel Time to Markets in Africa South of the Sahara, 2016](https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/YKDWJD) at 5-arc-minute (10km) resolution:

- `100` Low access (travel time greater than 4 hours)  
- `200` High access (travel time less than or equal to 4 hours)  

All layers were resampled to 1km and then we generated dominant classes across Ghana's districts (as household GPS locations are not available in any of the GLSS rounds).

```{r, fig.cap="Travel Time to Nearest Market by Size. Source: IFPRI/HarvestChoice, 2016."}

include_graphics("./fig/HC_tt10.png")

```

### Classification of Farm Holdings across Quadrants

Below is the resulting cross-classification summarized across Ghana's 170 districts (GLSS6) and 110 districts (GLSS5). Low and high market access districts use a **mean travel time to nearest 50K market**. In addition all districts that summarize to *unsuitable* `agpot-no` are reclassified to *low suitability* `agpot-lo`. None of the GLSS survey districts summarize to *low suitability*, the actual breakdown is given in \@ref(tab:gha-quadtab).

```{r gha-quadtab, results="asis"}

tmp <- gha[svyCode!="gha-glss4", .(District=uniqueN(paste(region, svyL2Nm)), N=.N),
  keyby=.(svyCode=as.character(svyCode), seg_quad_6)]
tmp1 <- fread("
svyCode,        seg_quad_6, District,  N
gha-glss6, agpot-lo / mkt-lo,        0,  0
gha-glss6, agpot-lo / mkt-hi,        0,  0
gha-glss5, agpot-lo / mkt-lo,        0,  0
gha-glss5, agpot-lo / mkt-hi,        0,  0")
tmp <- rbind(tmp, tmp1)

table_options(HTMLcaption="(#tab:gha-quadtab) Distribution of Ghana's Districts and Observations across Quadrants by Survey Year")
html(tabular((Survey=Factor(svyCode))*((Quadrant=seg_quad_6)+1)~((District=District)+(`Obs.`=N))*Heading()*Format(big.mark=",")*sum, data=tmp))

```

```{r gha-quadmap6, fig.cap="Quadrant Classification across Districts, Ghana 2012/13. Source: authors.", dev="png"}

print(spplot(g2[g2$svyCode=="gha-glss6",], "seg_mkt100k", col="grey20", lwd=.2, col.regions=viridis(2),
  xlab=list("Low/High Market Access\nacross Districts (100K, GLSS6)", cex=.9)),
  split=c(1,1,2,1), more=T)

print(spplot(g2[g2$svyCode=="gha-glss6",], "seg_quad_4", col="grey20", lwd=.2, col.regions=viridis(4),
  xlab=list("Quadrant Classification\nacross Districts (GLSS6)", cex=.9)),
  split=c(2,1,2,1), more=F)

```

```{r gha-quadmap5, fig.cap="Quadrant Classification across Districts, Ghana 2005/06. Source: authors.", dev="png"}

print(spplot(g2[g2$svyCode=="gha-glss5",], "seg_mkt100k", col="grey20", lwd=.2, col.regions=viridis(2),
  xlab=list("Low/High Market Access\nacross Districts (100K, GLSS5)", cex=.9)),
  split=c(1,1,2,1), more=T)

print(spplot(g2[g2$svyCode=="gha-glss5",], "seg_quad_4", col="grey20", lwd=.2, col.regions=viridis(4),
  xlab=list("Quadrant Classification\nacross Districts (GLSS5)", cex=.9)),
  split=c(2,1,2,1), more=F)

```


### Key Results across Quadrants

```{r gha-quad-fig2a, fig.cap="Est. Proportions of Farm Holdings across Quadrants (below/above 4 ha)"}

par(mfrow=c(1,3), mar=c(5,0,2,.25))
plot(svytable(~croparea_4ha+suit_bmgf, gha.svy[["gha6"]]), 
  main=NA, las=2, cex.axis=1, col=rev(viridis(3)),
  ylab=NA, xlab="Crop Suitability 3-class")
plot(svytable(~croparea_4ha+seg_mkt100k, gha.svy[["gha6"]]), 
  main=NA, las=2, cex.axis=1, col=viridis(2),
  ylab=NA, xlab="Market Access 2-class")
plot(svytable(~croparea_4ha+seg_quad_4, gha.svy[["gha6"]]), 
  main=NA, las=2, cex.axis=1, col=viridis(4)[c(1,2,3,4)],
  ylab=NA, xlab="Quadrant Typology 4-class")

```

```{r gha-quad-tab1, results="asis"}

tmp <- svymean(~I(interaction(suit_bmgf, seg_mkt100k)), gha.svy.shf[["gha6"]], na.rm=T)
tmp.cv <- confint(tmp)
tmp <- cbind(as.data.table(tmp*100), as.data.table(tmp.cv*100))
tmp[`2.5 %`<0, `2.5 %` := 0]
tmp <- tmp[, lapply(.SD, function(x) format(x, digits=0, nsmall=1, scientific=F))]
tmp[, CI := paste(`2.5 %`, `97.5 %`, sep=" - ")]

tmp[, `:=`(
  suit_bmgf = factor(rep(levels(gha$suit_bmgf), each=1, 2)),
  seg_mkt100k = factor(rep(levels(gha$seg_mkt100k), each=3, 1))
  )]

table_options(HTMLcaption="(#tab:tab1) Est. Proportions of Farm Holdings below 4 ha across Quadrants (2012/13, percent)")
html(tabular(Heading("Ag. Potential")*suit_bmgf*(mean+CI)~Heading("Market Access")*seg_mkt100k*Heading()*identity,
  data=tmp), rmarkdown=T)

```

Mean and median household characteristics (farms below 4 ha) across the resulting quadrants are estimated in the tables below.

```{r gha-quad-tab2, results="asis"}

tmp <- svyCrossTab(list(
  ~hhsize_imp, 
  ~agehead,
  ~I(100*femhead),
  ~I(100*widowhead),
  ~hhlabor,
  ~educhead,
  ~educave15_60,
  ~educhigh,
  ~I(100*ownhome), ~I(100*cellphone), ~I(100*telephone), ~I(100*electricity), 
  ~distwater, ~distroad, ~distpost, ~distbank, ~disthealth  
  ), ~seg_quad_4, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-quad-tab2) Est. Demographic Characteristics of Farm Holdings below 4 ha across Quadrants (2012/13, percent/km)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, nsmall=1, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

tmp <- svyCrossTab(list(
  ~landown,
  #~landshare,
  #~landrent,
  ~croparea_imp
  ), ~seg_quad_4, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-quad-tab3) Land Assets of Farm Holdings below 4 ha across Quadrants (2012/13, ha)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, nsmall=1, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

# Income
tmp <- svyCrossTab(list(
  ~aggross, 
  ~totgross, 
  ~I(100*naggross_sh),   
  ~cropvalue, 
  ~cropsales,
  ~I(100*cropsales_sh),    
  ~totlvstprod,
  ~totlivsold
  ), ~seg_quad_4, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-quad-tab4) Est. Production, Sales, and Income of Farm Holdings below 4 ha across Quadrants (2012/13, Percent/Cedis)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=1, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

# Livestock
tmp <- svyCrossTab(list(
~I(1E3*TLU_horse), ~I(1E3*TLU_cattle), ~I(1E3*TLU_pigs), ~I(1E3*TLU_sheep),
  ~I(1E3*TLU_small), ~I(1E3*TLU_total)), 
  ~seg_quad_4, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-quad-tab5) Est. Small and Large Livestock Holdings for Farms below 4 ha across Quadrants (2012/13, '000 TLU)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

# Input uses
tmp <- svyCrossTab(list(
  ~I(100*seeds), ~I(100*fert_any), ~I(100*fert_inorg), ~I(100*fert_org), 
  ~I(100*herb), ~I(100*pest), ~I(100*irr), ~I(100*fuel), ~I(100*hired_labor)), 
  ~seg_quad_4, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-quad-tab6) Est. Agricultural Input Use in Farms below 4 ha across Quadrants (2012/13, percent of farms)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

```

Below are estimated distributions of key characteristics for farms below 4 ha as of 2012/13.

```{r gha-quad-bplot2a, fig.cap="Est. Farm Size and Cultivated Area of Farm Holdings below 4 ha across Quadrants (2012/13, ha)"}

par(mfrow=c(1,2), cex=.8)
svyboxplot(landown~seg_quad_4, gha.svy.shf[["gha6"]], varwidth=T,
  ylab="Est. Land Owned (ha)", ylim=c(0,5))

svyboxplot(croparea_imp~seg_quad_4, gha.svy.shf[["gha6"]], varwidth=T,
  ylab="Est. Cultivated Area (ha)", ylim=c(0,5))

```

```{r gha-quad-bplot2c, fig.cap="Est. Gross Farm Income and Crop Sales for Farm Holdings below 4 ha across Quadrants (2012/13, '000 Cedis)"}

par(mfrow=c(1,2), cex=.8)
svyboxplot(I(aggross/1000)~seg_quad_4, gha.svy.shf[["gha6"]], varwidth=T,
  ylab="Est. Gross Farm Income ('000 Cedis)", ylim=c(0,5))

svyboxplot(I(cropsales/1000)~seg_quad_4, gha.svy.shf[["gha6"]], varwidth=T,
  ylab="Est. Crop Sales ('000 Cedis)", ylim=c(0,5))

```

```{r gha-quad-bplot2d, fig.cap="Est. Input Uses for Farm Holdings below 4 ha across Quadrants (2012/13, percent of farms)"}

par(mfrow=c(1,2), cex=.8)
svyboxplot(I(100*seeds)~seg_quad_4, gha.svy.shf[["gha6"]],
  ylab="Est. Seed Purchases (percent)", varwidth=T)

svyboxplot(I(100*fert_org)~seg_quad_4, gha.svy.shf[["gha6"]],
  ylab="Est. Organic Fertilizer Use (percent)",  varwidth=T)

svyboxplot(I(100*fert_inorg)~seg_quad_4, gha.svy.shf[["gha6"]],
  ylab="Est. Chemical Fertilizer Use (percent)", varwidth=T)

svyboxplot(I(100*fuel)~seg_quad_4, gha.svy.shf[["gha6"]],
  ylab="Est. Fuel/Maintenance Use (percent)", varwidth=T)

```


### Development Domain and Farm Types

In this section we combine results from the farm typology developed in [Section 3](./typology-of-farm-holdings.html) with the spatial dimensions outlined above (FAO/GAEZ agricultural suitability and market access using IFPRI/HarvestChoice travel time to nearest **100K** market).


```{r gha-class5-fig, fig.cap="Est. Proportions of Farm Holdings across Combined Categories (farms below 4 ha)"}

# Use `seg_quad_4` and a new 5-class typology `class5`
par(mar=c(3,.25,1,.25))
plot(svytable(~class5+seg_quad_4, gha.svy.shf[["gha6"]]), 
  main=NA, las=1, cex.axis=1, col=viridis(4),
  ylab=NA, xlab="Typology 5-class (below 4 ha)")

```

```{r gha-class5-tab, results="asis"}

tmp <- svymean(~I(interaction(seg_quad_4, class5)), gha.svy.shf[["gha6"]], na.rm=T)
tmp.cv <- confint(tmp)
tmp <- cbind(as.data.table(tmp*100), as.data.table(tmp.cv*100))
tmp[`2.5 %`<0, `2.5 %` := 0]
tmp <- tmp[, lapply(.SD, function(x) format(x, digits=0, nsmall=1, scientific=F))]
tmp[, CI := paste(`2.5 %`, `97.5 %`, sep=" - ")]

tmp[, `:=`(
  seg_quad_4 = factor(rep(levels(gha$seg_quad_4), each=1, 5),
    levels=levels(gha$seg_quad_4)),
  class5 = factor(rep(levels(gha$class5), each=4, 1), 
    levels=levels(gha$class5))
)]

table_options(HTMLcaption="(#tab:gha-tab-class5) Est. Proportions of Farm Holdings below 4 ha across Development Domains and Farm Types (2012/13, percent)")
html(tabular(Heading("Development Domain")*seg_quad_4*(mean+CI)~Heading("Farm Type")*class5*Heading()*identity,
  data=tmp), rmarkdown=T)

# TODO Need a function for column-wise proportions only


```

Mean and median household characteristics (farms below 4 ha) across the resulting quadrants are estimated in the tables below.

```{r gha-class5-tab2, results="asis"}

tmp <- svyCrossTab(list(
  ~hhsize_imp, 
  ~agehead,
  ~I(100*femhead),
  ~I(100*widowhead),
  ~hhlabor,
  ~educhead,
  ~educave15_60,
  ~educhigh,  
  ~I(100*ownhome), ~I(100*cellphone), ~I(100*telephone), ~I(100*electricity), 
  ~distwater, ~distroad, ~distpost, ~distbank, ~disthealth  
  ), ~class5, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-class5-tab2) Est. Demographic Characteristics of Farm Holdings below 4 ha across Farm Types (2012/13, percent/km)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, nsmall=1, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

tmp <- svyCrossTab(list(
  ~landown,
  ~landshare,
  ~landrent,
  ~croparea_imp
  ), ~class5, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-class5-tab3) Land Assets of Farm Holdings below 4 ha across Farm Types (2012/13, ha)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, nsmall=1, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

# Income
tmp <- svyCrossTab(list(
  ~aggross, 
  ~totgross, 
  ~I(100*naggross_sh),   
  ~cropvalue, 
  ~cropsales,
  ~I(100*cropsales_sh),    
  ~totlvstprod,
  ~totlivsold
  ), ~class5, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-class5-tab4) Est. Production, Sales, and Income of Farm Holdings below 4 ha across Farm Types (2012/13, Percent/Cedis)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=1, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

# Livestock
tmp <- svyCrossTab(list(
~I(1E3*TLU_horse), ~I(1E3*TLU_cattle), ~I(1E3*TLU_pigs), ~I(1E3*TLU_sheep),
  ~I(1E3*TLU_small), ~I(1E3*TLU_total)), 
  ~class5, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-class5-tab5) Est. Small and Large Livestock Holdings for Farms below 4 ha across Farm Types (2012/13, '000 TLU)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

# Input uses
tmp <- svyCrossTab(list(
  ~I(100*seeds), ~I(100*fert_any), ~I(100*fert_inorg), ~I(100*fert_org), 
  ~I(100*herb), ~I(100*pest), ~I(100*irr), ~I(100*fuel), ~I(100*hired_labor)), 
  ~class5, gha.svy.shf[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-class5-tab6) Est. Agricultural Input Use in Farms below 4 ha across Farm Types (2012/13, percent of farms)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

```

Below are estimated distributions of key characteristics for farms below 4 ha as of 2012/13.

```{r gha-class5-bplot2a, fig.cap="Est. Farm Size and Cultivated Area of Farm Holdings below 4 ha across Farm Types (2012/13, ha)"}

par(mfrow=c(1,2), cex=.8)
svyboxplot(landown~class5, gha.svy.shf[["gha6"]], varwidth=T,
  ylab="Est. Land Owned (ha)", ylim=c(0,5))

svyboxplot(croparea_imp~class5, gha.svy.shf[["gha6"]], varwidth=T,
  ylab="Est. Cultivated Area (ha)", ylim=c(0,5))

```

```{r gha-class5-bplot2c, fig.cap="Est. Gross Farm Income and Crop Sales for Farm Holdings below 4 ha across Farm Types (2012/13, '000 Cedis)"}

par(mfrow=c(1,2), cex=.8, las=2)
svyboxplot(I(aggross/1000)~class5, gha.svy.shf[["gha6"]], varwidth=T,
  ylab="Est. Gross Farm Income ('000 Cedis)", ylim=c(0,5))

svyboxplot(I(cropsales/1000)~class5, gha.svy.shf[["gha6"]], varwidth=T,
  ylab="Est. Crop Sales ('000 Cedis)", ylim=c(0,5))

```

```{r gha-class5-bplot2d, fig.cap="Est. Input Uses for Farm Holdings below 4 ha across Farm Types (2012/13, percent of farms)"}

par(mfrow=c(1,2), cex=.8)
svyboxplot(I(100*seeds)~class5, gha.svy.shf[["gha6"]],
  ylab="Est. Seed Purchases (percent)", varwidth=T)

svyboxplot(I(100*fert_org)~class5, gha.svy.shf[["gha6"]],
  ylab="Est. Organic Fertilizer Use (percent)",  varwidth=T)

svyboxplot(I(100*fert_inorg)~class5, gha.svy.shf[["gha6"]],
  ylab="Est. Chemical Fertilizer Use (percent)", varwidth=T)

svyboxplot(I(100*fuel)~class5, gha.svy.shf[["gha6"]],
  ylab="Est. Fuel/Maintenance Use (percent)", varwidth=T)

```

### Revised Development Domains

In this section we combine results from the farm typology developed in [Section 3](./typology-of-farm-holdings.html) with the spatial dimensions outlined above (agricultural suitability and market access using **[GLUES suitability](http://geoportal-glues.ufz.de/stories/globalsuitability.html)** product and IFPRI/HarvestChoice travel time to nearest **100K** market). GLUES suitability grid was modified to account for unsuitable forested, protected, and urban areas (white pixels).

```{r gha-glues, fig.cap="Areas of Low/High Agricultural Suitability (Ghana, 1980-2011)", dev="png", fig.width=7.6}

print(spplot(crop(glues, g2), col.regions=c("#ffffff", pal.ndvi(255)),
  xlab=list("Agricultural Suitability (GLUES 1980-2011, mod.)", cex=.9)),
  split=c(1,1,2,1), more=T)

print(spplot(g2[g2$svyCode=="gha-glss6",], "seg_suit_glues", col="grey70", lwd=.2, col.regions=pal.ndvi(4)[2:3],
  xlab=list("Low/High Suitability across Districts (GLSS6, GLUES mod.)", cex=.9)),
  split=c(2,1,2,1), more=F)

```

The 2 maps below show Ghana's districts (GLSS6) classified across areas of lo/hi ag. potential and lo/hi market access (using below/above **4hrs** travel time to nearest **100K** market).

```{r gha-mkt100k-fig, fig.cap="Spatial Quadrant Segmentation", dev="png", fig.width=7.6}

print(spplot(g2[g2$svyCode=="gha-glss6",], "seg_mkt100k", col="grey60", lwd=.2, col.regions=viridis(2),
  xlab=list("Low/High Market Access\nacross Districts (GLSS6, 100K)", cex=.9)),
  split=c(1,1,2,1), more=T)

print(spplot(g2[g2$svyCode=="gha-glss6",], "seg_quad_glues_4", col="grey60", lwd=.2, col.regions=viridis(4),
  xlab=list("Quadrant Classification\nacross Districts (GLSS6)", cex=.9)),
  split=c(2,1,2,1), more=F)

```

First we show the distribution of small/large farms across the revised quadrant typology.

```{r gha-class5glues-fig1, fig.cap="Est. Proportions of Farm Holdings across Quadrants (below/above 4 ha)"}

par(mfrow=c(1,3), mar=c(5,0,2,.25))
plot(svytable(~croparea_4ha+seg_suit_glues, gha.svy[["gha6"]]), 
  main=NA, las=2, cex.axis=1, col=pal.ndvi(4)[2:3],
  ylab=NA, xlab="GLUES Crop Suitability")
plot(svytable(~croparea_4ha+seg_mkt100k, gha.svy[["gha6"]]), 
  main=NA, las=2, cex.axis=1, col=viridis(2),
  ylab=NA, xlab="Market Access 2-class")
plot(svytable(~croparea_4ha+seg_quad_glues_4, gha.svy[["gha6"]]), 
  main=NA, las=2, cex.axis=1, col=viridis(4),
  ylab=NA, xlab="Quadrant Typology 4-class")

```

```{r gha-class5glues-tab1, results="asis"}

tmp <- svymean(~I(interaction(seg_suit_glues, seg_mkt100k)), gha.svy.shf[["gha6"]], na.rm=T)
tmp.cv <- confint(tmp)
tmp <- cbind(as.data.table(tmp*100), as.data.table(tmp.cv*100))
tmp[`2.5 %`<0, `2.5 %` := 0]
tmp <- tmp[, lapply(.SD, function(x) format(x, digits=0, nsmall=1, scientific=F))]
tmp[, CI := paste(`2.5 %`, `97.5 %`, sep=" - ")]

tmp[, `:=`(
  seg_suit_glues = factor(rep(levels(gha$seg_suit_glues), each=1, 2)),
  seg_mkt100k = factor(rep(levels(gha$seg_mkt100k), each=2, 1))
  )]

table_options(HTMLcaption="(#tab:gha-class5glues-tab1) Est. Proportions of Farm Holdings below 4 ha across Quadrants (2012/13, percent)")
html(tabular(Heading("Ag. Potential")*seg_suit_glues*(mean+CI)~Heading("Market Access")*seg_mkt100k*Heading()*identity,
  data=tmp), rmarkdown=T)

```

Below we show the distribution of small holdings across quadrants and farm types.

```{r gha-class5glues-fig2, fig.cap="Est. Proportions of Farm Holdings across Farm Types and Quadrants (farms below 4 ha, GLUES crop suitability)"}

# Use `seg_quad_glues_4` and 5-class typology `class5`
par(mar=c(3,.25,1,.25), cex=.9)
plot(svytable(~class5+seg_quad_glues_4, gha.svy.shf[["gha6"]]), 
  main=NA, las=1, cex.axis=1, col=viridis(4),
  ylab=NA, xlab="Typology 5-class across Quadrants (farms below 4 ha)")

```

```{r gha-class5glues-tab2, results="asis"}

# TODO Need a function for column-wise proportions only


tmp <- svymean(~I(interaction(seg_quad_glues_4, class5)), gha.svy.shf[["gha6"]], na.rm=T)
tmp.cv <- confint(tmp)
tmp <- cbind(as.data.table(tmp*100), as.data.table(tmp.cv*100))
tmp[`2.5 %`<0, `2.5 %` := 0]
tmp <- tmp[, lapply(.SD, function(x) format(x, digits=0, nsmall=1, scientific=F))]
tmp[, CI := paste(`2.5 %`, `97.5 %`, sep=" - ")]

tmp[, `:=`(
  seg_quad_glues_4 = factor(rep(levels(gha$seg_quad_glues_4), each=1, 5),
    levels=levels(gha$seg_quad_glues_4)),
  class5 = factor(rep(levels(gha$class5), each=4, 1), 
    levels=levels(gha$class5))
)]

table_options(HTMLcaption="(#tab:gha-class5glues-tab2) Est. Proportions of Farm Holdings below 4 ha across Development Domains and Farm Types (2012/13, percent)")
html(tabular(Heading("Development Domain (GLUES)")*seg_quad_glues_4*(mean+CI)~Heading("Farm Type")*class5*Heading()*identity,
  data=tmp), rmarkdown=T)

```
