==> R/05-clustering.Rmd <==
# Clustering Farm Holdings

## Nigeria

[placeholder]

## Ghana

```{r}

library(cluster)
library(mclust)
library(factoextra)
library(cowplot)
library(viridis)
#library(ggradar)

load("../tmp/2017-agra-aasr_GHA_cluster.RData")

# Training set for plotting
train <- gha[wave=="wave 6" & croparea_imp > 0]

```

```{r gha-clust, eval=FALSE}

library(data.table)
library(stringr)
library(survey)

setwd("~/Projects/2017-agra-aasr")
load("./tmp/2017-agra-aasr_GHA.RData")

# Preview this chapter only
setwd("~/Projects/2017-agra-aasr/R")
bookdown::preview_chapter("05-clustering.Rmd")

#####################################################################################
# Helper- Add alpha transparency
#####################################################################################
add.alpha <- function(col, alpha=1) {
  if(missing(col)) stop("Please provide a vector of colours.")
  apply(sapply(col, col2rgb)/255, 2, function(x) rgb(x[1], x[2], x[3], alpha=alpha))  
}

#####################################################################################
Y <- c("hhid", "croparea_4ha", "region", "svyL2Cd", "svyL2Nm", "agehead", "distbank")
  
#####################################################################################
# PAM Partition Clustering - 6 covariates
#####################################################################################
# Note: for large datasets `pam` can be very slow and `clara` is recommended.
# Sub-select variables (also always use standardized values)
# Select clustering covariates
X <- c("croparea_imp", "cropsales_sh", "naggross_sh", "totgross", "cropsales", "totcropprod")

# Select training set, keep all farm households from GLSS6
train <- gha[wave=="wave 6" & croparea_imp > 0, .SD, .SDcols=c(Y, X)]
train <- na.omit(train)
trainX <- data.frame(train)[, X]
trainX.std <- scale(trainX)

# Original approach (no dissimilarity matrix)
mod.pam.nb <- fviz_nbclust(trainX.std, pam, method="silhouette")
mod.pam.nb.wss <- fviz_nbclust(trainX.std, pam, method="wss")
plot(mod.pam.nb) # => suggests 3 and 5,6,7 equally
plot(mod.pam.nb.wss) # => elbows at 3 and 5

# Test Nb of clusters, using 3 distance measures (incl. Pearson collinearity)
# The Minkowski distance is a metric in a normed vector space which can be considered
# as a generalization of both the Euclidean distance and the Manhattan distance.
distE <- dist(trainX.std, method="euclidean")
distM <- dist(trainX.std, method="manhattan")
distP <- get_dist(trainX, method="pearson", stand=T)

# Compare stats
mod.pam.nbE <- fviz_nbclust(trainX, pam, method="silhouette", diss=distE, k.max=8)
mod.pam.nbE.wss <- fviz_nbclust(trainX, pam, method="wss", diss=distE, k.max=8)
mod.pam.nbM <- fviz_nbclust(trainX, pam, method="silhouette", diss=distM, k.max=8)
mod.pam.nbM.wss <- fviz_nbclust(trainX, pam, method="wss", diss=distM, k.max=8)
mod.pam.nbP <- fviz_nbclust(trainX, pam, method="silhouette", diss=distP, k.max=8)

plot(mod.pam.nbE) # => suggests 2, but 3,4,5 equally okay
plot(mod.pam.nbE.wss) # => elbow at 3 and 6
plot(mod.pam.nbM) # => suggests 2, but 3.4.5 esqually okay
plot(mod.pam.nbM.wss) # => elbow at 3 and 6
plot(mod.pam.nbP) # => worse
# => keep simple Euclidean distance for the rest of the analysis

# PAM clusters with 3 and 5 groups (5 is for Peter)
# VERY SLOW using eclust(), switched back to pam()
mod.pam3 <- pam(trainX, k=3, stand=T, metric="euclidean")
mod.pam5 <- pam(trainX, k=5, stand=T, metric="euclidean")

mod.pam3$medoids
#      croparea_imp cropsales_sh naggross_sh  totgross cropsales totcropprod
# [1,]     1.821085    0.3928416  0.07764422  1159.133       200    849.1333
# [2,]     1.214057    0.2622664  0.97018160 15344.550       120    457.5500
# [3,]     4.046856    0.6911727  0.19084414  4296.700      2263   3321.7000

mod.pam5$medoids
#     croparea_imp cropsales_sh naggross_sh   totgross cropsales totcropprod
# [1,]     1.982960    0.6403612  0.15914758  1885.0428       440   1010.0428
# [2,]     1.214057    0.0810154  0.01279659   625.1666        50    617.1666
# [3,]     1.214057    0.1424163  0.96412667 11744.0998        60    421.3000
# [4,]     4.405978    0.7028891  0.06977772  3869.4297      2530   3599.4297
# [5,]     2.428114    0.4020385  0.98563815 97159.3889       515   1349.3889

# Plot clusters
fviz_cluster(mod.pam5, geom="point", stand=F,
  ggtheme=theme_bw(base_size=10), palette=viridis(5))

# Save clusters
train[, clust_pam3 := mod.pam3$clustering]
train[, clust_pam5 := mod.pam5$clustering]
# => merge these vars into `gha` using code below

```

```{r gha-clust2, eval=FALSE}

#####################################################################################
# PAM - 2 covariates
#####################################################################################
# On Peter's suggestion
X <- c("cropsales_sh", "naggross_sh")
train <- gha[wave=="wave 6" & croparea_imp>0, .SD, .SDcols=c(Y, X)]
train <- na.omit(train)
trainX <- data.frame(train)[, X]

# Original approach (no dissimilarity matrix)
mod.pam.2var.nb <- fviz_nbclust(trainX, pam, method="silhouette")
mod.pam.2var.nb.wss <- fviz_nbclust(trainX, pam, method="wss")

plot(mod.pam.2var.nb) # => suggests 4
plot(mod.pam.2var.nb.wss) # => also 4 or 3

# PAM clusters with 5 groups
mod.pam5_2var <- pam(trainX, k=5, stand=F, metric="euclidean")

mod.pam5_2var$medoids
#      cropsales_sh naggross_sh
# [1,]   0.02463256  0.96299515
# [2,]   0.82927155  0.02321790
# [3,]   0.00000000  0.00000000
# [4,]   0.72440803  0.88793274
# [5,]   0.44375244  0.02282237

# Plot clusters
fviz_cluster(mod.pam5_2var, geom="point", stand=F,
  ggtheme=theme_bw(base_size=10), palette=viridis(5))

# Make pretty labels
mod.pam5_2var$clustering <- factor(mod.pam5_2var$clustering, 
  levels=c(3,5,2,1,4),
  labels=c("subs.", "pre-comm.", "specd. comm.", "trans.", "divf. comm."))

# Save clusters (and merge into `gha` below)
train[, clust_pam5_2var := mod.pam5_2var$clustering]

```

```{r gha-hclust, eval=FALSE}

#####################################################################################
# Hierarchical Clustering
#####################################################################################
# 0) Simple hclust()
mod.hc <- hclust(distE, method="median")
mod.hc <- hclust(distM, method="median")
mod.hc <- hclust(distP, method="median")

# 1) AGNES - Agglomerative Nesting
mod.agg <- eclust(trainX, "agnes", k.max=7, stand=T, hc_metric="euclidean")
mod.agg$ac
# [1] 0.9999284

# Cut tree into 5 groups and compare models
grp <- hcut(mod.agg, k=5)
table(grp)
train[, clust_agg := grp]

fviz_cluster(mod.agg, 
  ggtheme=theme_bw(base_size=10), palette=viridis(5), main="AGNES, k=5")

# 2) DIANA - DIvisive ANAlysis Clustering
mod.div <- eclust(trainX, "diana", k.max=7, stand=T, hc_metric="euclidean")
mod.div$dc
# [1] 0.9994045

# Cut tree into 5 groups and compare models
grp <- hcut(mod.div, k=5)
table(grp)
train[, clust_div := grp]

fviz_cluster(mod.div,
  ggtheme=theme_bw(base_size=10), palette=viridis(5), main="DIANA, k=5")

# 3) HKMEANS
mod.hk5 <- hkmeans(trainX, k=5, hc.method="ward.D2", stand=T, hc_metric="euclidean")
hkmeans_tree(mod.hk5, rect.col=viridis(5))

fviz_cluster(mod.hk5,
  ggtheme=theme_bw(base_size=10), palette=viridis(5), main="HKmeans, k=5")


#####################################################################################
# TODO Mclust (model-based clustering)
#####################################################################################

# Compare model fits using BIC
BIC <= mclustBIC(train, G=1:4)
plot(BIC)
summary(BIC)

# Cluster using BIC model selection
mod1 <- Mclust[train[, var], x=BIC]
summary(mod1, parameters=T)
plot(mod1, "classification")
table(mod1$classification)

d <- dist(train, method="euclidean")
fit <- hclust(d, method="ward.D")
plot(fit, hang=-1)
groups <- cutree(fit, k=4) # cut tree into 5 clusters
# draw dendogram with red borders around the 5 clusters 
rect.hclust(fit, k=4, border="red")

fit.any <- fit
fit.3 <- Mclust(tmp, G=3)
summary(fit.3) # display the best model

plot(fit.3, what="BIC")
plot(fit.3, what="classification")
plot(fit.3, what="density", type="image", col="dodgerblue3", grid=100)

table(fit.3$classification)
# -- with croparea_imp 
#   1    2    3 
# 695 3927 2827 
# -- without croparea_imp
#    1    2    3 
# 3155  681 3613 
# -- croparea_imp > 0
#    1    2    3 
# 3740  832 4539 

```

```{r gha-fpc, eval=FALSE}

library(fpc)

X <- c("cropsales_sh", "naggross_sh")
trainX <- data.frame(train)[, X]

# eps is reachability distance. Increase eps to include more points to clusters
# x is a matrix with 2 columns
bClus <- dbscan(trainX, 0.1) 
plot(bClus, trainX)
# => not working on these 2 vars
mod.pamk <- pamk(trainX, krange=2:7)
plotcluster(mod.pamk, train$hhid)
cluster.varstats

# Determine which cluster the observations within x2 belongs to
predict(dbClus, trainX)

# Try Mean Shift clustering
# https://cran.r-project.org/web/packages/MeanShift/vignettes/MeanShift-clustering.html
library(MeanShift)

# Select training set, keep all farm households from GLSS6
X <- c("croparea_imp", "cropsales_sh", "naggross_sh", "totgross", "cropsales", "totcropprod")
train <- gha[wave=="wave 6" & croparea_imp>0, .SD, .SDcols=c(Y, X)]
train <- na.omit(train)
trainX <- data.frame(train)[, X]
trainX.std <- scale(trainX)

h.cand <- quantile(dist(trainX), seq(.05, .40, .05))
model.ms <- lapply(h.cand, function(x) msClustering(trainX, h=x))
# Plot and vary h bandwidth
plot(trainX[,2], trainX[,3], col=viridis(7)[model.ms[[3]]$labels],
  main="Mean shift labels", cex=0.65, pch=16 )
plot(model.ms[[7]]$components[, "mode1"], model.ms[[3]]$components[, "mode2"], col=viridis(7)[model.ms[[2]]$labels],
  main="Mean shift labels", cex=0.65, pch=16 )


```

```{r gha-clust-save, eval=FALSE}

# Merge clustering results into `gha`, update survey designs, re-estimate population
setkey(train, hhid)
setkey(gha, hhid)

# Clean up
gha[, `:=`(
  clust_pam=NULL, clust_hca=NULL, clust_m=NULL, 
  clust_pam3=NULL, clust_pam5=NULL)]

gha[train, `:=`(clust_pam3=i.clust_pam3, clust_pam5=clust_pam5)]
gha[, .N, by=.(svyCode, clust_pam5)]
#      svyCode clust_pam5    N
# 1: gha-glss4         NA 5998
# 2: gha-glss5         NA 8687
# 3: gha-glss6         NA 9578
# 4: gha-glss6          1 2380
# 5: gha-glss6          2 1567
# 6: gha-glss6          3 1593
# 7: gha-glss6          4 1220
# 8: gha-glss6          5  434

# Add Peter's run (note that this is a different `train` dataset)
gha[, `:=`(clust_pam5_2var=NULL)]
gha[train, `:=`(clust_pam5_2var=i.clust_pam5_2var)]

gha[, .N, by=.(svyCode, clust_pam5_2var)]
#      svyCode clust_pam5_2var    N
# 1: gha-glss4              NA 5998
# 2: gha-glss5              NA 8687
# 3: gha-glss6              NA 9425
# 4: gha-glss6          trans. 1577
# 5: gha-glss6    specd. comm. 1515
# 6: gha-glss6           subs. 1286
# 7: gha-glss6     divf. comm. 1383
# 8: gha-glss6       pre-comm. 1586

gha.svy[["gha6"]] <- svydesign(~clust+hhid, strata=~region+rural, w=~weight, nest=T,
    data=gha[survey=="2012/13"])
gha.svy.farm <- lapply(gha.svy, subset, !is.na(croparea_clas))

# Save model results
save(add.alpha, X, pplot, svyCrossTab,
  mod.pam3, mod.pam5, mod.pam5_2var, 
  mod.pam.nb, mod.pam.nb.wss, mod.pam.2var.nb, mod.pam.2var.nb.wss,
  gha, gha.svy, gha.svy.farm,
  file="./tmp/2017-agra-aasr_GHA_cluster.RData")

```

We use a non-hierarchical **PAM** (Partitioning around Medoids) clustering approach on the set of 7 variables listed below and data from Ghana GLSS6 initially (all farm households). The selected variables are defined as:

- `croparea_imp` cultivated area (in fact operated area) in ha (imputed)
- `cropsales` value of sales from crops and crop byproducts (Cedis)
- `totgross` gross household income (Cedis)
- `totcropprod` value of crop production (Cedis)
- `naggross_sh` non-farm income as share of total gross income
- `cropsales_sh` crop sales as share of total crop production (in fact share of all ag products sold)
- `aggross` gross farm income (Cedis) -- not included in covariates

<u>Note on **k-mean** and **pam** techniques</u>: the use of *means* in *k-means* implies that clustering is highly sensitive to outliers. This can severely affects the assignment of observations to clusters. PAM (also known as k-medoids clustering) can provide a more robust algorithm in some instances.

<u>Note on violin plots</u>: all distributional plots below show **median** line in **red**, **mean** in **green**, and the **blue** region is the inferred **95% confidence interval** of the mean. At present pirate plots cannot be drawn for an entire population (using weights), simple boxplots are used instead showing median and IQ range.

### Aproach #1 -- PAM Partitioning

#### PAM - 6 Covariates

With the selected variables (standardized) the optimal number of clusters (based on average silhouette width and total within sum of square) is **3** or **6/7**. We choose to test results using **k=3** and **k=5** in the following run.

```{r, fig.cap="PAM Clustering of GLSS6 Households (6 covariates, Euclidean Distance)"}

plot_grid(
  mod.pam.nb + 
    geom_vline(aes(xintercept=3), color="grey50", linetype="dashed") +
    geom_vline(aes(xintercept=5), color="grey50", linetype="dashed") +
    theme_bw(base_size=10) + theme(panel.grid=element_line(linetype="dotted")),
  mod.pam.nb.wss + 
    geom_vline(aes(xintercept=3), color="grey50", linetype="dashed") +
    geom_vline(aes(xintercept=5), color="grey50", linetype="dashed") +
    theme_bw(base_size=10) + theme(panel.grid=element_line(linetype="dotted"))
)

```

```{r, fig.cap="Clusters of Farm Households (PAM, k=3, covariates=6)", dev="png"}

# Plot clusters
fviz_cluster(mod.pam3, geom="point", ggtheme=theme_bw(base_size=10), palette=viridis(3), main="PAM Clusters, k=3, covariates=6")

```

The second plot shows how farm households are distributed amongst the resulting 5 clusters along our 2 axes of interest (crop sales and non-farm income). 

```{r, fig.cap="Clusters of Farm Households (PAM, k=5, covariates=6)", dev="png"}

# Plot clusters
fviz_cluster(mod.pam5, geom="point", ggtheme=theme_bw(base_size=10), palette=viridis(5), main="PAM Clusters, k=5", covariates=6)

fviz_cluster(mod.pam5, geom="point", ggtheme=theme_bw(base_size=10), palette=viridis(5), main="PAM Clusters, k=5", covariates=6, 
  stand=F, choose.vars=c("cropsales_sh", "naggross_sh"))

```

Pair-wise scatter plots across all covariates (cultivated area, sales, and income variables) colored by cluster (k=5). In particular the second plot shows how the resulting 5 clusters are distributed amongst our 2 covariates of interest.

```{r, eval=F, fig.cap="Pair-wise scatter plots colored by cluster (2 covariates)", dev="png"}

clPairs(train[, .SD, .SDcols=setdiff(X, c("naggross_sh", "cropsales_sh"))],
  train$clust_pam5, 16, add.alpha(viridis(5), .5), CEX=.8)

```

Cluster medoids (centers) are listed below for k=3 and k=5.

```{r}

kable(mod.pam3$medoids, caption="Cluster medoids (k=3)", 
  format.args=list(big.mark=",", digits=1))
kable(mod.pam5$medoids, caption="Cluster medoids (k=5)",
  format.args=list(big.mark=",", digits=1))

```

The violin plots below show the distribution of all 6 covariates in the **sample** of farm households across the resulting **k=5** clusters. Actual population estimates are provided under section [Key Results](#pam---6-covariates-1).

```{r, fig.cap="Distribution of Household Characteristics across 5 Clusters (PAM)", dev="png"}

par(mfrow=c(1,2))
for (i in X[1:3]) pplot(as.formula(paste0(i, "~clust_pam5")), train, alpha=.02, ylab=NA, xlab=i,
  note=train[, lapply(.SD, function(x) 100*sum(is.na(x)|x==0)/.N), .SDcols=i])

for (i in X[4]) pplot(as.formula(paste0(i, "~clust_pam5")), train, alpha=.02, ylab=NA, xlab=i, ylim=c(0, 32000),
  note=train[, lapply(.SD, function(x) 100*sum(is.na(x)|x==0)/.N), .SDcols=i])

for (i in X[5:6]) pplot(as.formula(paste0(i, "~clust_pam5")), train, alpha=.02, ylab=NA, xlab=i, ylim=c(0, 8000),
  note=train[, lapply(.SD, function(x) 100*sum(is.na(x)|x==0)/.N), .SDcols=i])

```

#### PAM: 2 covariates

In the next scheme we only retain 2 covariates `naggross_sh` (share of non-farm income) and `cropsales_sh` (share of crop sales). 

```{r, fig.cap="PAM Clustering of GLSS6 Households (2 covariates, Euclidean Distance)"}

plot_grid(
  mod.pam.2var.nb + 
    geom_vline(aes(xintercept=5), color="grey50", linetype="dashed") +
    theme_bw(base_size=10) + theme(panel.grid=element_line(linetype="dotted")),
  mod.pam.2var.nb.wss + 
    geom_vline(aes(xintercept=5), color="grey50", linetype="dashed") +
    theme_bw(base_size=10) + theme(panel.grid=element_line(linetype="dotted"))
)

```

```{r, fig.cap="Clusters of Farm Households (PAM, k=5, covariates=2)", dev="png", fig.height=4, fig.width=5}

# Plot clusters (make it publication-ready)
fviz_cluster(mod.pam5_2var, stand=F, geom="point", 
  xlab="Crop Sales (share)", ylab="Non-Farm Income (share)",
  ggtheme=theme_bw(base_size=10), palette=viridis(5), 
  main="Farm Households (Ghana 2012/13, pam k=5)")

# ggsave("./out/MB/2017-agra-aasr-glss6_clust_pam5_2var.png",
#   units="in", width=6, height=5, family="Roboto Condensed")

```

Actual population statistics across the modeled 5 clusters are under section [Key Results](#pam---2-covariates).

### Aproach #2 -- Hierarchical Clustering

[in progress, not used]

This method is sensitive to the choice of dissimilarity measure (distance matrix). **Euclidean** distance is often preferred, however a **correlation-based** distance (with similar observations sharing features that are more highly correlated) may be used to identify household profiles/preferences. Further there are multiple generic types of hierarchical clustering algorithms:

- **Agglomerative** -- "bottom up" approach: each observation starts in its own cluster, and pairs of clusters are merged as one moves up the hierarchy (uses R `agnes()`).  
- **Divisive** -- "top down" approach: all observations start in one cluster, and splits are performed recursively as one moves down the hierarchy (uses R `diana()`).  
- **HKmean** -- a hybrid hierarchical k-means clustering for optimizing clustering outputs (uses R `hkmean()`).

We contrast the 3 approaches, cutting the resulting trees at 5 stems.

```{r gha-hclust-freq, eval=F, results="asis"}

# Cut tree into 4 groups and compare models
table_options(HTMLcaption="(#tab:gha-clust-hca) Frequency counts of Observations in Agg. and Div. Clusters (k=4)")
html(tabular((`Agglomerative Clusters`=Factor(clust_agg))+1~((`Divisive Clusters`=Factor(clust_div))+1)*Format(big.mark=",")*Heading(), 
  data=gha[svyCode=="gha-glss6"]), rmarkdown=T)

```

```{r gha-hclust-plot, eval=F, dev="png", fig.cap="Results from Agg. an Div. Hierarchical Clustering (k=4)"}

fviz_cluster(mod.hclust.agg,
  ggtheme=theme_bw(base_size=10), palette=viridis(4),
  main="AGNES, k=4")

fviz_cluster(mod.hclust.div,
  ggtheme=theme_bw(base_size=10), palette=viridis(4),
  main="DIANA, k=4")

```

Below are descriptive characteristics for the **sample** of farm households across the resulting tree branches.

```{r, eval=F, fig.cap="Distribution of Household Characteristics across 4 Clusters (AGNES, k=4)", dev="png"}

par(mfrow=c(1,3))
for (i in X) pplot(as.formula(paste0(i, "~clust_agg")), gha[svyCode=="glss6"], 
  alpha=.02, ylab=NA, xlab=i,
  note=gha[svyCode=="glss6", lapply(.SD, function(x) 100*sum(is.na(x)|x==0)/.N), .SDcols=i])

```

```{r, eval=F, fig.cap="Distribution of Household Characteristics across 4 Clusters (DIANA, k=4)", dev="png"}

par(mfrow=c(1,3))
for (i in X) pplot(as.formula(paste0(i, "~clust_div")), gha[svyCode=="glss6"], 
  alpha=.02, ylab=NA, xlab=i,
  note=gha[svyCode=="glss6", lapply(.SD, function(x) 100*sum(is.na(x)|x==0)/.N), .SDcols=i])

```


### Key Results

#### PAM - 6 Covariates

Below are population summaries for the selected k=5 cluster scheme.


```{r gha-clust-radar, eval=F}

# Estimate mean/median/quantiles
tmp <- svyCrossTab(list(
  ~croparea_imp,
  ~aggross, 
  ~totgross, 
  ~I(100*naggross_sh),   
  ~cropvalue, 
  ~cropsales,
  ~I(100*cropsales_sh),    
  ~totlvstprod,
  ~totlivsold
  ), ~clust_pam5, gha.svy.farm[["gha6"]], quantiles=c(.25,.5,.75))

# Draw radar plots


ggradar(tmp) 

# ggplot(dt.group[varLabel %like% "value added"],
#   aes(year, value, color=varLabel)) +
#   geom_line() + xlab("") + ylab("") +
#   facet_wrap(~Region, ncol=2) +
#   scale_color_manual(name="",
#     values=c("darkgreen", "orange", "red", "brown"),
#     labels=str_replace(levels(dt.group$varLabel)[3:6], fixed("("), "\n(")) +
#   theme_bw(base_size=10) + 
#   theme(legend.position="right", panel.grid=element_line(linetype="dotted"))
# 
# ggsave("../out/MB/2017-agra-aasr_WDI_ts_gdp-sector.png", width=6, height=7, units="in")


```

Tables of summary statistics for the entire population. Proportions compared to manually-derived 5-class farm types.

```{r gha-clust-fig, fig.cap="Est. Proportions of Farm Holdings across Clusters and Farm Types (k=5, all farm households)"}

par(mar=c(2.2,0,0,0))
plot(svytable(~class5+clust_pam5, gha.svy.farm[["gha6"]]), 
  main=NA, las=1, cex.axis=1, col=viridis(5),
  ylab=NA, xlab="Farm Types across Clusters (k=5, farms below 4 ha)")

```

```{r gha-clust-tab1, results="asis"}

tmp <- svymean(~factor(clust_pam5), gha.svy.farm[["gha6"]], na.rm=T)
tmp.cv <- confint(tmp)
tmp <- cbind(as.data.table(tmp*100), as.data.table(tmp.cv*100))
tmp[`2.5 %`<0, `2.5 %` := 0]
tmp <- tmp[, lapply(.SD, function(x) format(x, digits=0, nsmall=1, scientific=F))]
tmp[, CI := paste(`2.5 %`, `97.5 %`, sep=" - ")]
tmp[, clust_pam5 := paste0("clust", 1:5)]

table_options(HTMLcaption="(#tab:gha-clust-tab1) Est. Proportions of Farm Holdings across Clusters (k=5, 2012/13, percent)")
html(tabular((`Cluster`=Factor(clust_pam5))*(mean+CI)~Heading("Proportion of Farm Households")*identity,
  data=tmp), rmarkdown=T)

```

Mean and median household characteristics across clusters are estimated in the tables below.

```{r gha-clust-tab3, results="asis"}

tmp <- svyCrossTab(list(
  ~hhsize_imp, 
  ~agehead,
  ~I(100*femhead),
  ~I(100*widowhead),
  ~hhlabor,
  ~educhead,
  ~educave15_60,
  ~educhigh,  
  ~I(100*ownhome), ~I(100*cellphone), ~I(100*telephone), ~I(100*electricity), 
  ~distwater, ~distroad, ~distpost, ~distbank, ~disthealth  
  ), ~factor(clust_pam5, labels=paste0("clust", 1:5)), gha.svy.farm[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-clust-tab3) Est. Demographic Characteristics of Farm Holdings across Clusters (2012/13, percent/km)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, nsmall=1, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

```

```{r gha-clust-tab2, results="asis"}

# Estimate mean/median/quantiles
tmp <- svyCrossTab(list(
  ~croparea_imp,
  ~aggross, 
  ~totgross, 
  ~I(100*naggross_sh),   
  ~cropvalue, 
  ~cropsales,
  ~I(100*cropsales_sh),    
  ~totlvstprod,
  ~totlivsold
  ), ~factor(clust_pam5, labels=paste0("clust", 1:5)), gha.svy.farm[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-clust-tab2) Estimated Mean and Median Characteristics across Clusters")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, nsmall=0, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

```

#### PAM - 2 Covariates

Below are population summaries for the selected k=5 cluster scheme (with only 2 covariates). First we show how farm types are distributed across small/large farm sizes (below/above 4 ha).

```{r, fig.cap=""}

par(mar=c(3,.25,.25,.25))
plot(svytable(~croparea_4ha+clust_pam5_2var, gha.svy.farm[["gha6"]]), 
  main=NA, las=1, cex.axis=1, col=viridis(5),
  ylab=NA, xlab="Farm Types (below/above 4 ha)")

```

```{r gha-clust2-tab1, results="asis"}

tmp <- svymean(~~I(interaction(clust_pam5_2var, croparea_4ha)), gha.svy.farm[["gha6"]], na.rm=T)
tmp.cv <- confint(tmp)
tmp <- cbind(as.data.table(tmp*100), as.data.table(tmp.cv*100))
tmp[`2.5 %`<0, `2.5 %` := 0]
tmp <- tmp[, lapply(.SD, function(x) format(x, digits=0, nsmall=1, scientific=F))]
tmp[, CI := paste(`2.5 %`, `97.5 %`, sep=" - ")]
tmp[, clust_pam5_2var := levels(gha$clust_pam5_2var)]
tmp[, clust_pam5_2var := factor(clust_pam5_2var, levels(gha$clust_pam5_2var))]

tmp[, `:=`(
  clust_pam5_2var = factor(rep(levels(gha$clust_pam5_2var), each=1, 2),
    levels=levels(gha$clust_pam5_2var)),
  croparea_4ha = factor(rep(levels(gha$croparea_4ha), each=5, 1), 
    levels=levels(gha$croparea_4ha))
)]

table_options(HTMLcaption="(#tab:gha-clust2-tab1) Est. Proportions of Farm Holdings across Clusters (k=5, 2012/13, percent)")
html(tabular((`Farm Type (cluster)`=clust_pam5_2var)*(mean+CI)~Heading("Proportion of Farm Households")*croparea_4ha*Heading()*identity,
  data=tmp), rmarkdown=T)

```

Mean and median household characteristics across farm types are estimated below.

```{r gha-clust2-tab3, results="asis"}

tmp <- svyCrossTab(list(
  ~hhsize_imp, 
  ~agehead,
  ~I(100*femhead),
  ~I(100*widowhead),
  ~hhlabor,
  ~educhead,
  ~educave15_60,
  ~educhigh,  
  ~I(100*ownhome), ~I(100*cellphone), ~I(100*telephone), ~I(100*electricity), 
  ~distwater, ~distroad, ~distpost, ~distbank, ~disthealth  
  ), ~clust_pam5_2var, gha.svy.farm[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-clust2-tab3) Est. Demographic Characteristics of Farm Holdings across Clusters (2012/13, percent/km)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, nsmall=1, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

```

```{r gha-clust2-tab2, results="asis"}

# Estimate mean/median
tmp <- svyCrossTab(list(
  ~croparea_imp,
  ~aggross, 
  ~totgross, 
  ~I(100*naggross_sh),   
  ~cropvalue, 
  ~cropsales,
  ~I(100*cropsales_sh),    
  ~totlvstprod,
  ~totlivsold
  ), ~clust_pam5_2var, gha.svy.farm[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-clust2-tab2) Estimated Mean and Median Characteristics across Clusters")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, nsmall=0, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

# Input uses
tmp <- svyCrossTab(list(
  ~I(100*seeds), ~I(100*fert_any), ~I(100*fert_inorg), ~I(100*fert_org), 
  ~I(100*herb), ~I(100*pest), ~I(100*irr), ~I(100*fuel), ~I(100*hired_labor)), 
  ~clust_pam5_2var, gha.svy.farm[["gha6"]], quantiles=c(.25,.5,.75))

table_options(HTMLcaption="(#tab:gha-clust2-tab6) Est. Agricultural Input Use across Farm Types (2012/13, percent of farms)")
html(tabular(Variable*(Mean+Q50)~Heading()*By*Heading()*Stat*Heading()*identity*Format(digits=0, big.mark=",", scientific=F),
  data=tmp), rmarkdown=T)

```
