==> R/06-macro_trends.Rmd <==
# Trends in Macro Indicators

```{r, include=FALSE}

library(data.table)
library(ggplot2)
library(stringr)
library(viridis)
library(knitr)

load("../tmp/2017-agra-aasr_WDI.RData")
# rmarkdown::render("./R/06-macro_trends.Rmd", "word_document", output_dir="./out/MB")

```


```{r wdi, eval=FALSE}

library(WDI)
library(data.table)
library(ggplot2)
library(stringr)
library(viridis)

setwd("~/Projects/2017-agra-aasr")
load("./tmp/2017-agra-aasr_WDI.RData")

#####################################################################################
# 2017.06.15 WDI Time Series for Peter
#####################################################################################
# We need:
# - cereal yields in kg/ha for SSA, with/without ZAF, China, South Asia (not China) for 2000-2015
# - GDP per capita for SSA, with/without ZAF
# - GDP shares by sector (ag, services, industry) SSA with/without ZAF
# WDI variables are:
# - AG.YLD.CREL.KG
# - NY.GDP.PCAP.PP.CD
# - NV.AGR.TOTL.ZS
# - NV.SRV.TETC.ZS
# - NV.IND.TOTL.ZS

# SSF ZG Sub-Saharan Africa (all income levels) Aggregates
# SSA ZF Sub-Saharan Africa (developing only) Aggregates
# ECS Z7 Europe & Central Asia (all income levels) Aggregates
# ECA 7E Europe & Central Asia (developing only) Aggregates
# EAS Z4 East Asia & Pacific (all income levels) Aggregates
# EAP 4E East Asia & Pacific (developing only) Aggregates

reg <- WDI_data[[2]]
reg <- data.table(reg)
reg[region %like% "Sub-Saharan Africa", .N, keyby=country]

# We need mean cereal yield weighted by cereal production I assume
vars <- c("AG.YLD.CREL.KG", "NY.GDP.PCAP.PP.CD",
  "NV.AGR.TOTL.ZS", "NV.SRV.TETC.ZS", "NV.IND.TOTL.ZS")

dt <- WDI(indicator=vars, country=c("ZG", "ZF", "Z4", "4E", "CN"), start=1990, end=2015)
dt <- data.table(dt)
dt <- melt(dt, id.vars=1:3)

levels(dt$variable)
dt[, varLabel := factor(variable, labels=c(
  "Cereal Yield\n(kg/ha)",
  "GDP per Capita\n(PPP, current int. $)",
  "Agriculture\n(value added, % of GDP)",
  "Services\n(valude added, % of GDP)",
  "Industry\n(value added, % of GDP)"))]

setnames(dt, "country", "Region")
dt[, Region := factor(Region)]

ggplot(dt[year<2015 & Region=="Sub-Saharan Africa"], aes(year, value, color=Region)) + geom_line() +
  xlab("") + ylab("") +
  facet_grid(varLabel ~ ., scales="free_y")

ggsave("./out/MB/2017-agra-aasr_WDI_SSA_ts.png", width=7, heigh=9, units="in")

ggplot(dt[year<2015 & iso2c=="ZF"], aes(year, value, color=Region)) + geom_line() +
  xlab("") + ylab("") +
  facet_grid(varLabel ~ ., scales="free_y")

ggsave("./out/MB/2017-agra-aasr_WDI_ZF_ts.png", width=7, heigh=9, units="in")

ggplot(dt[year<2015], aes(year, value, color=Region)) + geom_line() +
  xlab("") + ylab("") +
  facet_grid(varLabel ~ ., scales="free_y")

ggsave("./out/MB/2017-agra-aasr_WDI_ts.png", width=7, heigh=9, units="in")

#####################################################################################
# => noticed that only diff between region `SSA` and `SSA (excluded high income)` is
# Equatorial Guinea as ZAF is not included in High Income.
# So we need to regenerate regional aggregates by hand, using the correct formulas:
# - yield = sum(yield*production)/sum(production)
# - GPD per capita = sum(GDP)/sum(population)
# - value_added_% = 100*sum(value_added)/sum(GDP)
# Also need to exclude `China` from `East Asia`

# Generate new regional groupings
reg[region %like% "Sub-Saharan Africa", unique(iso3c), by=income]
reg[region %like% "East Asia", unique(country), by=.(income, iso3c)]
reg[region %like% "East Asia" & !income %like% "High", unique(country), by=income]
reg[region %like% "South Asia" & !income %like% "High", unique(country), by=.(income, iso3c)]

reg[, group := NULL]
reg[region %like% "Sub-Saharan Africa" & !income %like% "High" & iso3c != "ZAF",
  group := "Rest of sub-Saharan Africa (excl. high income)"]
reg[region %like% "East Asia" & !income %like% "High" & iso3c != "CHN",
  group := "Rest of East Asia (excl. high income)"]
reg[region %like% "South Asia" & !income %like% "High" & iso3c != "IND",
  group := "Rest of South Asia (excl. high income)"]

reg[iso3c=="ZAF", group := "South Africa"]
reg[iso3c=="CHN", group := "China"]
reg[iso3c=="IND", group := "India"]

reg[!is.na(group), unique(country), by=group]

# Re-download all needed WDI indicators
vars <- c("AG.YLD.CREL.KG", "AG.PRD.CREL.MT",
  #"NY.GDP.PCAP.PP.CD",
  "NY.GDP.MKTP.PP.CD", "SP.POP.TOTL",
  #"NV.AGR.TOTL.ZS", "NV.SRV.TETC.ZS", "NV.IND.TOTL.ZS", "NV.IND.MANF.ZS",
  "NV.AGR.TOTL.KD", "NV.SRV.TETC.KD", "NV.IND.TOTL.KD", "NV.IND.MANF.KD", "NY.GDP.MKTP.KD",
  #"EA.PRD.AGRI.KD",
  "AG.LND.AGRI.K2", "SL.AGR.EMPL.ZS", "SL.TLF.TOTL.IN")

WDIsearch(vars)
dt = WDI(indicator=vars, country=reg[!is.na(group), unique(iso2c)], start=1990, end=2015)
dt <- data.table(dt)

# Merge in group names
setkey(reg, iso2c)
setkey(dt, iso2c)
dt[reg, Region := group]

# Re-compute regional aggregates
# Make sure to only include full records (hence ifelse())
dt.group <- dt[, .(
  AG.YLD.CREL.KG =  sum(AG.YLD.CREL.KG*AG.PRD.CREL.MT, na.rm=T)/sum(ifelse(is.na(AG.YLD.CREL.KG), NA, AG.PRD.CREL.MT), na.rm=T),
  NY.GDP.PCAP.PP.CD = sum(ifelse(is.na(SP.POP.TOTL), NA, NY.GDP.MKTP.PP.CD), na.rm=T)/sum(ifelse(is.na(NY.GDP.MKTP.PP.CD), NA, SP.POP.TOTL), na.rm=T),
  NV.AGR.TOTL.ZS = 100*sum(ifelse(is.na(NY.GDP.MKTP.KD), NA, NV.AGR.TOTL.KD), na.rm=T)/sum(ifelse(is.na(NV.AGR.TOTL.KD), NA, NY.GDP.MKTP.KD), na.rm=T),
  NV.SRV.TETC.ZS = 100*sum(ifelse(is.na(NY.GDP.MKTP.KD), NA, NV.SRV.TETC.KD), na.rm=T)/sum(ifelse(is.na(NV.SRV.TETC.KD), NA, NY.GDP.MKTP.KD), na.rm=T),
  NV.IND.TOTL.ZS = 100*sum(ifelse(is.na(NY.GDP.MKTP.KD), NA, NV.IND.TOTL.KD), na.rm=T)/sum(ifelse(is.na(NV.IND.TOTL.KD), NA, NY.GDP.MKTP.KD), na.rm=T),
  NV.IND.MANF.ZS = 100*sum(ifelse(is.na(NY.GDP.MKTP.KD), NA, NV.IND.MANF.KD), na.rm=T)/sum(ifelse(is.na(NV.IND.MANF.KD), NA, NY.GDP.MKTP.KD), na.rm=T),  
  lABOR.AG.PROD = sum(NV.AGR.TOTL.KD, na.rm=T)/sum(ifelse(is.na(NV.AGR.TOTL.KD), NA, (SL.AGR.EMPL.ZS/100)*SL.TLF.TOTL.IN), na.rm=T),
  lAND.AG.PROD = sum(NV.AGR.TOTL.KD, na.rm=T)/sum(ifelse(is.na(NV.AGR.TOTL.KD), NA, AG.LND.AGRI.K2), na.rm=T)
), by=.(Region, year)]

# Replace Inf with NA
dt.group <- dt.group[, lapply(.SD, function(x) ifelse(x==Inf, NA, x))]

# Impute missing by country, by indicator using linear interpolation
setorder(dt.group, Region, year)
dt.group.imp <- dt.group[, lapply(.SD, zoo::na.approx, na.rm=F), .SDcols=-1, by=Region]

dt.group <- melt(dt.group.imp, id.vars=1:2)
levels(dt.group$variable)
dt.group[, varLabel := factor(variable, labels=c(
  "Cereal Yield (kg/ha)",
  "GDP per Capita (PPP, current int. $)",
  "Agriculture (value added, % of GDP)",
  "Services (value added, % of GDP)",
  "Industry (value added, % of GDP)",
  "Manufacturing (value added, % of GDP)",
  "Agricultural Value Added per Worker",
  "Agricultural Value Added per Hectare"))]

dt.group[, Region := factor(Region, levels=c(
  "South Africa", "Rest of sub-Saharan Africa (excl. high income)",
  "China", "Rest of East Asia (excl. high income)",
  "India", "Rest of South Asia (excl. high income)"))]

# Ruttanograms
dt.rutt <- dcast(dt.group[year %between% c(2000, 2014) & variable %like% ".AG.PROD"], Region+year~variable)
setnames(dt.rutt, 3:4, c("labor", "land"))
setorder(dt.rutt, Region, year)
dt.rutt[, yearLabel := NULL]
dt.rutt[!is.na(land) & !is.na(labor), yearLabel := ifelse(year==min(year) | year==max(year), year, NA), by=Region]

# Save data
fwrite(setnames(setcolorder(dt, c(1,2,16,3:15)), "Region", "region"), "./out/MB/2017-agra-aasr_WDI_ts (corrected).csv")

# Save workspace
save.image("./tmp/2017-agra-aasr_WDI.RData")

```

Long-term time series of macro-economic indicators across regions. Note that these series are aggregated using WDI country-level indicators. Table \@ref(tab:agg-regions) below lists all countries included in each region. The raw data is available at [2017-agra-aasr_WDI_ts (corrected).csv](https://github.com/mbacou/2017-agra-aasr/blob/master/out/MB/2017-agra-aasr_WDI_ts%20(corrected).csv).


```{r agg-gdp, fig.cap="Regional Trends in GDP per Capita (1990-2015) Source: WDI.", fig.height=3}

ggplot(dt.group[varLabel %in% levels(varLabel)[2]],
  aes(year, value, color=Region)) +
  geom_line() + xlab("") + ylab("") +
  facet_wrap(~varLabel, nrow=1, scales="free_y") +
  scale_color_manual(name="", values=viridis_pal()(6),
    labels=str_replace(levels(dt.group$Region), fixed(" ("), "\n(")) +
  theme_bw(base_size=10) + 
  theme(legend.position="right", panel.grid=element_line(linetype="dotted"))

ggsave("../out/MB/2017-agra-aasr_WDI_ts_agg-gdp.png", width=6, height=4, units="in")

```

```{r agg-yield, fig.cap="Regional Trends in Cereal Yield (1990-2015) Source: WDI.", fig.height=3}

ggplot(dt.group[varLabel %in% levels(varLabel)[1]],
  aes(year, value, color=Region)) +
  geom_line() + xlab("") + ylab("") +
  facet_wrap(~varLabel, nrow=1, scales="free_y") +
  scale_color_manual(name="", values=viridis_pal()(6),
    labels=str_replace(levels(dt.group$Region), fixed(" ("), "\n(")) +
  theme_bw(base_size=10) + 
  theme(legend.position="right", panel.grid=element_line(linetype="dotted"))

ggsave("../out/MB/2017-agra-aasr_WDI_ts_agg-yield.png", width=6, height=4, units="in")

```


```{r agg-gdpsector, fig.cap="Regional Trends in Sectoral Value Added (1990-2015) Source: WDI.", fig.height=6, fig.width=6}

ggplot(dt.group[varLabel %like% "value added"],
  aes(year, value, color=varLabel)) +
  geom_line() + xlab("") + ylab("") +
  facet_wrap(~Region, ncol=2) +
  scale_color_manual(name="",
    values=c("darkgreen", "orange", "red", "brown"),
    labels=str_replace(levels(dt.group$varLabel)[3:6], fixed("("), "\n(")) +
  theme_bw(base_size=10) + 
  theme(legend.position="right", panel.grid=element_line(linetype="dotted"))

ggsave("../out/MB/2017-agra-aasr_WDI_ts_gdp-sector.png", width=6, height=7, units="in")

```

Note that India is missing multiple years of agricultural employment statistics. Simple **linear interpolation** is used to impute these missing values. This is a rough first-stab.

```{r agg-rutt, fig.cap="Trends in Land and Labor Productivity (2000-2014). Source: WDI.", fig.height=4, fig.width=6}

ggplot(dt.rutt,
  aes(labor/1000, land/1000, color=Region, group=Region, label=yearLabel)) +
  geom_point(size=.5) + geom_label(size=2, nudge_x=1, nudge_y=6, show.legend=F) +
  geom_path(na.rm=T) + ylim(0,170) +
  xlab("Ag. Value Added per Ag. Worker ('000 constant 2000 US$)") +
  ylab("Ag. Value Added per Hectare ('000 constant 2000 US$)") +
  scale_color_manual(name="",
    labels=str_replace(levels(dt.group$Region), fixed("("), "\n("),
    values=viridis_pal()(6)) +
  theme_bw(base_size=10) + theme(legend.position="right")

# ggsave("../docs/fig/2017-agra-aasr_RUTT_ts (corrected).svg", width=6, heigh=4, units="in")
# Make corrections in Inkscape

```

The regional aggregation used in the above graphs is provided here.

```{r agg-regions}

kable(reg[!is.na(group), .(`ISO2 Code`=iso2c, `ISO3 Code`=iso3c, `Income Level`=income), keyby=.(Region=group, Country=country)],
  caption="Regional Aggregation")

```
